% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{ac_latent.brmsfit}
\alias{ac_latent.brmsfit}
\alias{ac_latent}
\title{Extract Latent Autocorrelated Effects}
\usage{
\method{ac_latent}{brmsfit}(object, summary = TRUE, robust = FALSE, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}.}

\item{summary}{Should summary statistics be returned
instead of the raw values? Default is \code{TRUE}.}

\item{robust}{If \code{FALSE} (the default) the mean is used as
the measure of central tendency and the standard deviation as
the measure of variability. If \code{TRUE}, the median and the
median absolute deviation (MAD) are applied instead.
Only used if \code{summary} is \code{TRUE}.}

\item{probs}{The percentiles to be computed by the \code{quantile}
function. Only used if \code{summary} is \code{TRUE}.}

\item{...}{Currently ignored}
}
\value{
A list of arrays (one for each grouping level).
}
\description{
This function extracts the latent ARMA-correlated residuals
from a multi-level model of class \code{brmsfit}.
}
\examples{
\dontrun{
fit <- brm(riders ~ temp + windspeed + ar(1, gr = workday, latent = TRUE),
           data = bike_share_daily, family = gaussian(), chains = 2)
ac_latent(fit)
}

}
