% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/loo-helpers.R
\name{add_loo}
\alias{add_loo}
\alias{add_loo.brmsfit}
\title{Add the LOO information criterion to fitted model objects}
\usage{
add_loo(x, ...)

\method{add_loo}{brmsfit}(x, ...)
}
\arguments{
\item{x}{An \R object typically of class \code{brmsfit}.}

\item{...}{Further arguments passed to 
\code{\link[brms:LOO]{LOO}} or \code{\link[brms:WAIC]{WAIC}}.}
}
\value{
An object of the same class as \code{x}, but
  with the LOO information criterion added for later usage.
}
\description{
Add the LOO information criterion to fitted model objects
}
\details{
For more details see \code{\link[brms:add_ic]{add_ic}}.
}
