% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_call.R
\name{all.equal.brm_call}
\alias{all.equal.brm_call}
\title{Checks if two brm_call objects are equal}
\usage{
\method{all}{equal.brm_call}(target, current, ...)
}
\arguments{
\item{...}{Currently ignored}

\item{a}{A \code{brm_call} object}

\item{b}{Another \code{brm_call} object}
}
\description{
Checks if two brm_call objects are equal
}
