% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{as.brmsprior}
\alias{as.brmsprior}
\title{Transform into a brmsprior object}
\usage{
as.brmsprior(x)
}
\arguments{
\item{x}{An object to be transformed.}
}
\value{
A \code{brmsprior} object if the transformation was possible.
}
\description{
Try to transform an object into a \code{brmsprior} object.
}
