% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocor.R
\name{autocor.brmsfit}
\alias{autocor.brmsfit}
\alias{autocor}
\title{(Deprecated) Extract Autocorrelation Objects}
\usage{
\method{autocor}{brmsfit}(object, resp = NULL, ...)

autocor(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}.}

\item{resp}{Optional names of response variables. If specified, predictions
are performed only for the specified response variables.}

\item{...}{Currently unused.}
}
\value{
A \code{cor_brms} object or a list of such objects for multivariate
  models. Not supported for models fitted with brms 2.11.1 or higher.
}
\description{
(Deprecated) Extract Autocorrelation Objects
}
