% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{autocor}
\alias{autocor}
\title{Extract Autocorrelation Structures}
\usage{
autocor(x, ...)
}
\arguments{
\item{x}{A \R object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
What exactly is returned depends on the specific method.
}
\description{
Extract Autocorrelation Structures
}
