% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/generics.R
\name{bayes_factor.brmsfit}
\alias{bayes_factor.brmsfit}
\alias{bayes_factor}
\title{Bayes Factors from Marginal Likelihoods}
\usage{
\method{bayes_factor}{brmsfit}(x1, x2, log = FALSE, ...)

bayes_factor(x1, x2, ...)
}
\arguments{
\item{x1}{An \R object}

\item{x2}{Another \R object}

\item{log}{Report Bayes factors on the log-scale?}

\item{...}{Additional arguments passed to 
\code{\link[brms:bridge_sampler]{bridge_sampler}}.}
}
\description{
Generic function that computes Bayes factors from marginal likelihoods.
}
\details{
More details are provided under
  \code{\link[bridgesampling:bf]{bf}}.
}
\note{
The \code{bayes_factor} method is an alias of the
 \code{\link[bridgesampling:bf]{bf}} method provided by
 the \pkg{bridge_sampler} package. Using an alias
 is necessary, because the function name \code{bf}
 is already taken in \pkg{brms}.
}
\seealso{
\code{
  \link[brms:bridge_sampler]{bridge_sampler},
  \link[brms:post_prob]{post_prob}
}
}
