% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{bf}
\alias{bf}
\title{Set up a model formula for use in the \pkg{brms} package}
\usage{
bf(formula, ..., nonlinear = NULL)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): 
a symbolic description of the model to be fitted. 
The details of model specification are given under 'Details'.}

\item{...}{Additional \code{formula} objects to specify 
predictors of special model parts and auxiliary parameters. 
Formulas can either be named directly or contain
names on their left-hand side. Currently, the following
names corresponding are accepted: 
\code{sigma} (residual standard deviation of
the \code{gaussian} and \code{student} families);
\code{shape} (shape parameter of the \code{Gamma},
\code{weibull}, \code{negbinomial} and related
zero-inflated / hurdle families); \code{nu}
(degrees of freedom parameter of the \code{student} family);
\code{phi} (precision parameter of the \code{beta} 
and \code{zero_inflated_beta} families).}

\item{nonlinear}{An optional list of formuluas, specifying 
linear models for non-linear parameters. If \code{NULL} (the default)
\code{formula} is treated as an ordinary formula. 
If not \code{NULL}, \code{formula} is treated as a non-linear model
and \code{nonlinear} should contain a formula for each non-linear 
parameter, which has the parameter on the left hand side and its
linear predictor on the right hand side.
Alternatively, it can be a single formula with all non-linear
parameters on the left hand side (separated by a \code{+}) and a
common linear predictor on the right hand side.
More information is given under 'Details'.}
}
\description{
Set up a model formula for use in the \pkg{brms} package
allowing to define additive multilevel models for all parameters
of the assumed distribution of the response.
}

