% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{bridge_sampler.brmsfit}
\alias{bridge_sampler.brmsfit}
\alias{bridge_sampler}
\title{Log Marginal Likelihood via Bridge Sampling}
\usage{
\method{bridge_sampler}{brmsfit}(samples, ...)
}
\arguments{
\item{samples}{A \code{brmsfit} object.}

\item{...}{Additional arguments passed to 
\code{\link[bridgesampling:bridge_sampler]{bridge_sampler.stanfit}}.}
}
\description{
Computes log marginal likelihood via bridge sampling,
which can be used in the computation of bayes factors
and posterior model probabilities.
The \code{brmsfit} method is just a thin wrapper around
the corresponding method for \code{stanfit} objects.
}
\details{
More details are provided under
  \code{\link[bridgesampling:bridge_sampler]{bridge_sampler}}.
}
\seealso{
\code{
  \link[brms:bayes_factor]{bayes_factor},
  \link[brms:post_prob]{post_prob}
}
}
