% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cache.R
\name{brmsfit_needs_refit}
\alias{brmsfit_needs_refit}
\title{Check whether a given cached fit can be used without refitting.}
\usage{
brmsfit_needs_refit(
  cached_fit,
  sdata,
  scode,
  algorithm,
  silent = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{cached_fit}{old fit (e.g. loaded from file)}

\item{sdata}{new Stan data (result of a call to \code{\link{make_standata}}).
pass \code{NULL} to avoid data check.}

\item{scode}{new Stan code (result of a call to \code{\link{make_stancode}}).
pass \code{NULL} to avoid code check.}

\item{algorithm}{new algorithm. Pass \code{NULL} to avoid algorithm check.}

\item{silent}{if \code{TRUE}, no messages will be given}

\item{verbose}{if \code{TRUE} detailed report of the differences is printed
to the console.}
}
\value{
a boolean indicating whether a refit is needed
}
\description{
Use with \code{verbose = TRUE} to get additional info on how the stored
fit differs from the given data and code.
}
