% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlations.R
\name{cor_ar}
\alias{cor.ar}
\alias{cor_ar}
\title{AR(p) correlation structure}
\usage{
cor_ar(formula = ~1, p = 1, cov = FALSE)
}
\arguments{
\item{formula}{A one sided formula of the form ~ t, or ~ t | g,
specifying a time covariate t and, optionally, a grouping factor g.
A covariate for this correlation structure must be integer valued.
When a grouping factor is present in \code{formula}, the correlation structure
is assumed to apply only to observations within the same grouping level;
observations with different grouping levels are assumed to be uncorrelated.
Defaults to ~ 1, which corresponds to using the order of the observations
in the data as a covariate, and no groups.}

\item{p}{A non-negative integer specifying the autoregressive (AR) order of the ARMA structure.
Default is 0.}

\item{cov}{A flag indicating whether ARMA effects should be estimated by means
  of residual covariance matrices
  (currently only possible for stationary ARMA effects of order 1).
  If \code{FALSE} (the default) a regression formulation
  is used that is considerably faster and allows for ARMA effects
  of order higher than 1 but cannot handle user defined standard errors.}
}
\value{
An object of class \code{cor_arma} containing solely autoregression terms.
}
\description{
This function is a constructor for the \code{cor_arma} class, allowing for autoregression terms only.
}
\details{
As of \pkg{brms} version 0.6.0, the AR structure refers to autoregressive effects of residuals
  to match the naming and implementation in other packages such as nlme.
  Previously, the AR term in \pkg{brms} referred to autoregressive effects of the response.
  The latter are now named ARR effects and can be modeled using argument \code{r} in the
  \code{cor_arma} and \code{cor_arr} functions.
}
\examples{
cor_ar(~visit|patient, p = 2)
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}
\seealso{
\code{\link{cor_arma}}
}

