% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cor_car}
\alias{cor_car}
\alias{cor_icar}
\title{Spatial conditional autoregressive (CAR) structures}
\usage{
cor_car(W, formula = ~1, type = c("escar", "esicar"))

cor_icar(W, formula = ~1)
}
\arguments{
\item{W}{Adjacency matrix}

\item{formula}{An optional one-sided formula of the form 
\code{~ 1 | g}, where \code{g} is a grouping factor mapping
observations to spatial locations. If not specified,
each observation is treated as a separate location.}

\item{type}{Type of the CAR structure. Currenlty implemented
are \code{"escar"} (exact sparse CAR) and \code{"esicar"}
(exact sparse intrinsic CAR). More information is
provided in the 'Details' section.}
}
\description{
These functions are constructors for the \code{cor_car} class
implementing spatial conditional autoregressive structures.
}
