% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-predictor.R
\name{data_predictor}
\alias{data_predictor}
\title{Prepare Predictor Data}
\usage{
data_predictor(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A named list of data related to predictor variables.
}
\description{
Prepare data related to predictor variables in \pkg{brms}.
Only exported for use in package development.
}
\keyword{internal}
