% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-response.R
\name{data_response}
\alias{data_response}
\title{Prepare Response Data}
\usage{
data_response(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A named list of data related to response variables.
}
\description{
Prepare data related to response variables in \pkg{brms}.
Only exported for use in package development.
}
\keyword{internal}
