% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Dirichlet}
\alias{Dirichlet}
\alias{ddirichlet}
\alias{rdirichlet}
\title{The Dirichlet Distribution}
\usage{
ddirichlet(x, alpha, log = FALSE)

rdirichlet(n, alpha)
}
\arguments{
\item{x}{Matrix of quantiles. Each row corresponds to one probability vector.}

\item{alpha}{Matrix of positive shape parameters. Each row corresponds to one
probability vector.}

\item{log}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{n}{Number of draws to sample from the distribution.}
}
\description{
Density function and random number generation for the dirichlet
distribution with shape parameter vector \code{alpha}.
}
\details{
See \code{vignette("brms_families")} for details on the
parameterization.
}
