% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{DiscreteWeibull}
\alias{DiscreteWeibull}
\alias{ddiscrete_weibull}
\alias{pdiscrete_weibull}
\alias{qdiscrete_weibull}
\alias{rdiscrete_weibull}
\title{The Discrete Weibull Distribution}
\usage{
ddiscrete_weibull(x, mu, shape, log = FALSE)

pdiscrete_weibull(x, mu, shape, lower.tail = TRUE, log.p = FALSE)

qdiscrete_weibull(p, mu, shape, lower.tail = TRUE, log.p = FALSE)

rdiscrete_weibull(n, mu, shape)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{mu}{Location parameter in the unit interval.}

\item{shape}{Positive shape parameter.}

\item{log}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{lower.tail}{Logical; If \code{TRUE} (default), return P(X <= x). 
Else, return P(X > x) .}

\item{log.p}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{p}{Vector of probabilities.}

\item{n}{Number of samples to draw from the distribution.}
}
\description{
Density, distribution function, quantile function and random generation 
for the discrete Weibull distribution with location \code{mu} and
shape \code{shape}.
}
\details{
See \code{vignette("brms_families")} for details
on the parameterization.
}
