% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{.brms_digest}
\alias{.brms_digest}
\title{Internal helper: digest wrapper used by all hash methods}
\usage{
.brms_digest(object, algo = "xxhash64")
}
\description{
Internal helper: digest wrapper used by all hash methods
}
\keyword{internal}
