% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{draws-brms}
\alias{draws-brms}
\alias{as_draws.brmsfit}
\alias{as_draws_matrix.brmsfit}
\alias{as_draws_array.brmsfit}
\alias{as_draws_df.brmsfit}
\alias{as_draws_list.brmsfit}
\alias{as_draws_rvars.brmsfit}
\title{Transform \code{brmsfit} to \code{draws} objects}
\usage{
\method{as_draws}{brmsfit}(x, ...)

\method{as_draws_matrix}{brmsfit}(x, ...)

\method{as_draws_array}{brmsfit}(x, ...)

\method{as_draws_df}{brmsfit}(x, ...)

\method{as_draws_list}{brmsfit}(x, ...)

\method{as_draws_rvars}{brmsfit}(x, ...)
}
\arguments{
\item{x}{A \code{brmsfit} object or another \R object for which
the methods are defined.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\description{
Transform a \code{brmsfit} object to a format supported by the 
\pkg{posterior} package.
}
\examples{
\dontrun{
fit <- brm(count ~ zAge + zBase * Trt + (1|patient),
           data = epilepsy, family = poisson())
           
# extract posterior draws in an array format
(draws_fit <- as_draws_array(fit))
posterior::summarize_draws(draws_fit)

# extract posterior draws in a random variables format
as_draws_rvars(fit)
}

}
\seealso{
\code{\link[posterior:draws]{draws}}
}
