% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ensure_type}
\alias{ensure_type}
\title{Ensure an object satisfies a type check}
\usage{
ensure_type(obj, pred, arg = deparse(substitute(obj)), doc = "NULL")
}
\arguments{
\item{obj}{The object to check.}

\item{pred}{A function (like is.data.frame or inherits) or a string (like "character").}

\item{arg}{Optional argument name for clearer error messages.}
}
\value{
The object, invisibly (for piping).
}
\description{
Ensure an object satisfies a type check
}
