% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/generics.R
\name{expose_functions.brmsfit}
\alias{expose_functions.brmsfit}
\alias{expose_functions}
\title{Expose user-defined \pkg{Stan} functions}
\usage{
\method{expose_functions}{brmsfit}(x, ...)

expose_functions(x, ...)
}
\arguments{
\item{x}{An \R object}

\item{...}{Further arguments}
}
\description{
Export user-defined \pkg{Stan} function to the 
\code{\link[base:environment]{.GlobalEnv}}.
For more details see 
\code{\link[rstan:expose_stan_functions]{expose_stan_functions}}.
}
