% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{family.brmsfit}
\alias{family.brmsfit}
\title{Extract Model Family Objects}
\usage{
\method{family}{brmsfit}(object, resp = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}.}

\item{resp}{Optional names of response variables. If specified, predictions
are performed only for the specified response variables.}

\item{...}{Currently unused.}
}
\value{
A \code{brmsfamily} object
or a list of such objects for multivariate models.
}
\description{
Extract Model Family Objects
}
