% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-helpers.R
\name{get_cache_folder}
\alias{get_cache_folder}
\title{Determine Cache Folder for brms}
\usage{
get_cache_folder(file)
}
\arguments{
\item{file}{A file path (string) that may include a directory component.}
}
\value{
A string indicating which folder to use for caching.
}
\description{
Checks if the provided file path includes a directory. If so, returns that directory.
Otherwise, looks for a user‐defined `brms_cache_folder` option. If the option is not set,
defaults to the current directory (".").
}
