% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-helpers.R
\name{get_dpar}
\alias{get_dpar}
\title{Get samples of a distributional parameter from a prepared object.}
\usage{
get_dpar(prep, dpar, i = NULL, ilink = NULL)
}
\arguments{
\item{prep}{a 'brmsprep' or 'mvbrmsprep' object}

\item{dpar}{name of the distributional parameter}

\item{i}{the current observation number}

\item{ilink}{should the inverse link function be applied?
if NULL the value is chosen internally}
}
\value{
If the parameter is predicted and \code{i} is NULL or 
  \code{length(i) > 1}, an \code{S x N} matrix.
  If the parameter it not predicted or \code{length(i) == 1},
  a vector of length \code{S}. Here \code{S} is the number of samples
  and \code{N} is the number of observations or 
  length of \code{i} if specified.
}
\description{
Get samples of a distributional parameter from a \code{brmsprep} or 
\code{mvbrmsprep} object. This function is primarily useful when developing
custom families or packages depending on \pkg{brms}. 
This function lets callers easily handle both the case when the
distributional parameter is predicted directly, via a linear
predictor or fixed to a constant. See the vignette
 \code{vignette("brms_customfamilies")} for an example usage.
}
\examples{
\dontrun{
  posterior_predict_my_dist <- function(i, prep, ...) {
    mu <- brms::get_dpar(prep, "mu", i = i)
    mypar <- brms::get_dpar(prep, "mypar", i = i)
    my_rng(mu, mypar)
  }
} 
}
