% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{get_inits.brmsfit}
\alias{get_inits.brmsfit}
\title{Extract Initial Values Used for Each Chain}
\usage{
\method{get_inits}{brmsfit}(object, ...)
}
\arguments{
\item{object}{A brmsfit object.}

\item{...}{Currently ignored.}
}
\value{
A list of initial values for each chain (if available).
}
\description{
Extract the initial values used by Stan for each chain.
This is available if the model was fitted with
\code{rstan} backend.
}
