% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-syntax-helpers.R
\name{get_sp_vars}
\alias{get_sp_vars}
\title{store_uni_me <- function(x, ...) {
  # store unique names of noise-free terms in all 'me' elements
  UseMethod("store_uni_me")
}}
\usage{
get_sp_vars(x, type)
}
\description{
#' @export
store_uni_me.mvbrmsterms <- function(x, uni_me = NULL, ...) {
  if (is.null(uni_me)) {
    uni_me <- get_uni_me(x)
  }
  for (i in seq_along(x$terms)) {
    x$terms[[i]] <- store_uni_me(x$terms[[i]], uni_me = uni_me, ...)
  }
  x
}
}
\details{
#' @export
store_uni_me.brmsterms <- function(x, uni_me = NULL, ...) {
  if (is.null(uni_me)) {
    uni_me <- get_uni_me(x)
  }
  if (!length(uni_me)) {
    return(x)
  }
  for (i in seq_along(x$dpars)) {
    x$dpars[[i]] <- store_uni_me(x$dpars[[i]], uni_me = uni_me, ...)
  }
  x
}

#' @export
store_uni_me.btnl <- function(x, uni_me = NULL, ...) {
  if (is.null(uni_me)) {
    uni_me <- get_uni_me(x)
  }
  for (i in seq_along(x$nlpars)) {
    x$nlpars[[i]] <- store_uni_me(x$nlpars[[i]], uni_me = uni_me, ...)
  }
  x
}

#' @export
store_uni_me.btl <- function(x, uni_me = NULL, ...) {
  if (is.null(uni_me)) {
    uni_me <- get_uni_me(x)
  }
  if (length(uni_me)) {
    attr(x[["sp"]], "uni_me") <- uni_me
  }
  x
}
}
