% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{get_y}
\alias{get_y}
\title{Extract response values}
\usage{
get_y(x, resp = NULL, warn = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{brmsfit}} object.}

\item{resp}{Optional names of response variables for which to extract values.}

\item{warn}{For internal use only.}

\item{...}{Further arguments passed to \code{\link{standata}}.}
}
\value{
Returns a vector of response values for univariate models and a
  matrix of response values with one column per response variable for
  multivariate models.
}
\description{
Extract response values from a \code{\link{brmsfit}} object.
}
\keyword{internal}
