% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-helpers.R
\name{gp}
\alias{gp}
\title{Set up gaussian process terms in \pkg{brms}}
\usage{
gp(...)
}
\arguments{
\item{...}{One or more terms containing covariates for
the gaussian process.}
}
\description{
Function used to set up a gaussian process term in \pkg{brms}.
The function does not evaluate its arguments --
it exists purely to help set up a model with gaussian process terms.
}
\seealso{
\code{\link[brms:brmsformula]{brmsformula}}
}
