% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_mixture.R
\name{hu_families}
\alias{hu_families}
\title{List of all hurdle families, where "hurdle" means anything where observed 
zeros are guaranteed to be due to the zero-inflation component.}
\usage{
hu_families()
}
\description{
List of all hurdle families, where "hurdle" means anything where observed 
zeros are guaranteed to be due to the zero-inflation component.
}
