% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{inits.brmsfit}
\alias{inits.brmsfit}
\alias{inits}
\title{Extract Initial Values Used for Each Chain}
\usage{
\method{inits}{brmsfit}(x, ...)

inits(x, ...)
}
\arguments{
\item{x}{A \code{brmsfit} object.}

\item{...}{Currently ignored.}
}
\value{
The initial values (either user-specified or generated randomly) for
 all chains. This is a list with one component per chain. Each component is
 a named list containing the initial values for each parameter for the
 corresponding chain.
}
\description{
Extract the initial values used by Stan for each chain.
This is currently only available if the model was fitted with
the \code{rstan} backend.
}
