% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_logit_m1_to_half}
\alias{inv_logit_m1_to_half}
\title{Modified inverse logit-link taking values between -1 and 0.5}
\usage{
inv_logit_m1_to_half(x)
}
\arguments{
\item{x}{A numeric or complex vector.}
}
\description{
Computes \code{inv_logit(y) * 1.5 - 1}
}

