% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_call.R
\name{is.brm_call}
\alias{is.brm_call}
\title{Checks if argument is a \code{brm_call} object}
\usage{
is.brm_call(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{brm_call} object
}
