% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-class.R
\name{is.brmsfit}
\alias{is.brmsfit}
\title{Checks if argument is a \code{brmsfit} object}
\usage{
is.brmsfit(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{brmsfit} object
}
