% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-class.R
\name{is.brmsfit_multiple}
\alias{is.brmsfit_multiple}
\title{Checks if argument is a \code{brmsfit_multiple} object}
\usage{
is.brmsfit_multiple(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{brmsfit_multiple} object
}
