% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsformula.R
\name{is.brmsformula}
\alias{is.brmsformula}
\title{Checks if argument is a \code{brmsformula} object}
\usage{
is.brmsformula(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{brmsformula} object
}
