% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{is.brmsprior}
\alias{is.brmsprior}
\title{Checks if argument is a \code{brmsprior} object}
\usage{
is.brmsprior(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{brmsprior} object
}
