% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsterms.R
\name{is.brmsterms}
\alias{is.brmsterms}
\title{Checks if argument is a \code{brmsterms} object}
\usage{
is.brmsterms(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{brmsterms} object
}
\seealso{
\code{\link[brms:parse_bf]{parse_bf}}
}
