% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsformula.R
\name{is.mvbrmsformula}
\alias{is.mvbrmsformula}
\title{Checks if argument is a \code{mvbrmsformula} object}
\usage{
is.mvbrmsformula(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{mvbrmsformula} object
}
