% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is.mvbrmsterms}
\alias{is.mvbrmsterms}
\title{Checks if argument is a \code{mvbrmsterms} object}
\usage{
is.mvbrmsterms(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{mvbrmsterms} object
}
\seealso{
\code{\link[brms:parse_bf]{parse_bf}}
}
