% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.generics.R
\name{linear.predictor}
\alias{linear.predictor}
\title{Compute linear predictors term}
\usage{
linear.predictor(x, ...)
}
\arguments{
\item{x}{A fitted model object typically of class \code{brmsfit}.}

\item{...}{Currently ignored}
}
\value{
Usually, an S x N matrix containing the linear predictor samples, where S is the number of samples
  and N is the number of observations in the data.
}
\description{
Compute linear predictors term
}

