% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.generics.R
\name{linear_predictor}
\alias{linear_predictor}
\alias{linear_predictor.brmsfit}
\title{Compute linear predictors term}
\usage{
linear_predictor(x, ...)
}
\arguments{
\item{x}{A fitted model object typically of class \code{brmsfit}.}

\item{...}{Currently ignored}
}
\value{
Usually, an S x N matrix containing the linear predictor samples, where S is the number of samples
  and N is the number of observations in the data.
}
\description{
Compute linear predictors term
}
\examples{
\dontrun{
 fit_i2 <- brm(rating ~ treat + period + carry + (1|subject),
             data = inhaler, family = "gaussian")
 eta <- linear_predictor(fit_i2)
 }
}

