% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logit_m1_to_half}
\alias{logit_m1_to_half}
\title{Modified logit-link returning values between -1 and 0.5}
\usage{
logit_m1_to_half(x)
}
\arguments{
\item{x}{A numeric or complex vector.}
}
\description{
Computes \code{logit((x + 1) / 1.5)}
}

