% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.generics.R
\name{loglik}
\alias{loglik}
\alias{loglik.brmsfit}
\title{Compute the pointwise log-likelihood}
\usage{
loglik(x, ...)
}
\arguments{
\item{x}{A fitted model object typically of class \code{brmsfit}.}

\item{...}{Currently ignored}
}
\value{
Usually, an S x N matrix containing the pointwise log-likelihood samples, where S is the number of samples
  and N is the number of observations in the data.
}
\description{
Compute the pointwise log-likelihood
}

