% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric-helpers.R
\name{logm1}
\alias{logm1}
\title{Logarithm with a minus one offset.}
\usage{
logm1(x, base = exp(1))
}
\arguments{
\item{x}{A numeric or complex vector.}

\item{base}{A positive or complex number: the base with respect to which
logarithms are computed. Defaults to \emph{e} = \code{exp(1)}.}
}
\description{
Computes \code{log(x - 1)}.
}
