% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-helpers.R
\name{loo_compare.brmsfit}
\alias{loo_compare.brmsfit}
\alias{loo_compare}
\title{Model comparison with the \pkg{loo} package}
\usage{
\method{loo_compare}{brmsfit}(x, ..., criterion = c("loo", "waic",
  "kfold"), model_names = NULL)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{More fitted model objects or further arguments
passed to the underlying post-processing functions.}

\item{criterion}{The name of the criterion to be extracted 
from \code{brmsfit} objects.}

\item{model_names}{If \code{NULL} (the default) will use model names 
derived from deparsing the call. Otherwise will use the passed 
values as model names.}
}
\value{
An object of class "\code{compare.loo}".
}
\description{
Model comparison with the \pkg{loo} package
}
\details{
See \code{\link[loo:loo_compare]{loo::loo_compare}} for details.
}
\examples{
\dontrun{
# model with population-level effects only
fit1 <- brm(rating ~ treat + period + carry,
            data = inhaler)
add_criterion(fit1) <- "waic"

# model with an additional varying intercept for subjects
fit2 <- brm(rating ~ treat + period + carry + (1|subject),
            data = inhaler)
add_criterion(fit2) <- "waic"

# compare both models
loo_compare(fit1, fit2, criterion = "waic")
}

}
