% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-helpers.R
\name{loo_compare}
\alias{loo_compare}
\title{Model comparison with the \pkg{loo} package}
\arguments{
\item{x}{An object of class "\code{loo}" or a list of such objects.}

\item{...}{Additional objects of class "\code{loo}".}
}
\value{
An object of class "\code{compare.loo}".
}
\description{
Model comparison with the \pkg{loo} package
}
\details{
See \code{\link[loo:loo_compare]{loo::loo_compare}} for details.
}
\examples{
\dontrun{
# model with population-level effects only
fit1 <- brm(rating ~ treat + period + carry,
            data = inhaler)
(waic1 <- waic(fit1))

# model with an additional varying intercept for subjects
fit2 <- brm(rating ~ treat + period + carry + (1|subject),
            data = inhaler)
(waic2 <- waic(fit2))

# compare both models
(loo_compare(waic1, waic2))
}

}
