% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_moment_match.R
\name{loo_moment_match.brmsfit}
\alias{loo_moment_match.brmsfit}
\alias{loo_moment_match}
\title{Moment matching for efficient approximate leave-one-out cross-validation}
\usage{
\method{loo_moment_match}{brmsfit}(x, loo, ...)
}
\arguments{
\item{x}{An object of class \code{brmsfit}.}

\item{loo}{An object of class \code{loo} originally created from \code{x}.}

\item{...}{Further arguments passed to the underlying methods.
Additional arguments initially passed to \code{\link{loo}},
for example, \code{newdata} or \code{resp} need to be passed
again to \code{loo_moment_match} in order for the latter
to work correctly.}
}
\value{
An updated object of class \code{loo}.
}
\description{
Moment matching for efficient approximate leave-one-out cross-validation 
(LOO-CV). See \code{\link[loo:loo_moment_match]{loo_moment_match}} 
for more details.
}
\details{
The moment matching algorithm requires samples 
  of all variables defined in Stan's \code{parameters} block
  to be saved. Otherwise \code{loo_moment_match} cannot be computed.
  Thus, please set \code{save_all_pars = TRUE} in the call to \code{brm},
  if you are planning to apply \code{loo_moment_match} to your models.
}
\examples{
\dontrun{
fit1 <- brm(count ~ zAge + zBase * Trt + (1|patient),
            data = epilepsy, family = poisson(),
            save_all_pars = TRUE)
            
# throws warning about some pareto k estimates being too high
(loo1 <- loo(fit1))
(mmloo1 <- loo_moment_match(fit1, loo = loo1))
}

}
