% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects-helpers.R
\name{make_conditions}
\alias{make_conditions}
\title{Prepare Fully Crossed Conditions}
\usage{
make_conditions(x, vars, show_vars = FALSE, sep = " | ", ...)
}
\arguments{
\item{x}{An \R object from which to extract the variables
that should be part of the conditions.}

\item{vars}{Names of the variables that should be part of the conditions.}

\item{show_vars}{Logical; indicates if variables names should be 
part of the rownames of the output. Defaults to \code{FALSE}.}

\item{sep}{A single character string defining the separator
used in the rownames of the output.}

\item{...}{Currently unused.}
}
\value{
A \code{data.frame} where each row indicates a condition.
}
\description{
This is a helper function to prepare fully crossed conditions primarily 
for use with the \code{conditions} argument of \code{\link{marginal_effects}}.
}
\details{
For factor like variables, all levels are used as conditions.
  For numeric variables, \code{mean + (-1:1) * SD} are used as conditions.
}
\examples{
df <- data.frame(x = c("a", "b"), y = rnorm(10))
make_conditions(df, vars = c("x", "y"))

}
\seealso{
\code{\link{marginal_effects}}
}
