% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/generics.R
\name{margins_plot.brmsfit}
\alias{margins_plot}
\alias{margins_plot.brmsfit}
\title{Display marginal effects of predictors}
\usage{
\method{margins_plot}{brmsfit}(x, effects = NULL, data = NULL,
  re_formula = NA, ncol = NULL, method = c("fitted", "predict"),
  rug = FALSE, theme = "gray", do_plot = TRUE, ...)

margins_plot(x, ...)
}
\arguments{
\item{x}{An object usually of class \code{brmsfit}}

\item{effects}{An optional character vector naming effects
(main effects or interactions) for which to compute marginal plots.
If \code{NULL} (the default), plots for all effects are generated.}

\item{data}{An optional \code{data.frame} containing variable values
to marginalize on. Each effect defined in \code{effects} will
be plotted for each row of \code{data}. It is recommended
to only define a few rows in order to keep the plots clear.
If \code{NULL} (the default), numeric variables will be marginalized
by using their means and factors will get their reference level assigned.}

\item{re_formula}{A formula containing random effects to be considered 
in the marginal predictions. If \code{NULL}, include all random effects; 
if \code{NA} (default), include no random effects.}

\item{ncol}{Number of plots to display per column for each effect.
If \code{NULL} (default), \code{ncol} is computed internally based
on the number of rows of \code{data}.}

\item{method}{Either \code{"fitted"} or \code{"predict"}. 
If \code{"fitted"}, plot marginal predictions of the regression curve. 
If \code{"predict"}, plot marginal predictions of the responses.}

\item{rug}{Logical; indicating whether a rug representation of predictor
values should be added via \code{\link[ggplot2:geom_rug]{geom_rug}}.
Default is \code{FALSE}.}

\item{theme}{The ggplot theme to use. For details see
\code{\link[ggplot2:ggtheme]{ggtheme}}.}

\item{do_plot}{logical; indicates if plots should be
plotted directly in the active graphic device.
Defaults to \code{TRUE}.}

\item{...}{Currently ignored.}
}
\value{
A list of ggplot objects one for each effect.
}
\description{
Display marginal effects of one or more numeric and/or categorical 
predictors including interaction effects of order 2.
}
\examples{
\dontrun{
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c + (1 | patient),
           data = epilepsy, family = poisson()) 
## plot all marginal effects
margins_plot(fit)
## only plot the marginal interaction effect of 'log_Base4_c:Trt_c'
## for different values for 'log_Age_c'
mdata <- data.frame(log_Age_c = c(-0.3, 0, 0.3))
margins_plot(fit, effects = "log_Base4_c:Trt_c", 
             data = mdata)
## also incorporate random effects variance over patients
## and add a rug representation of predictor values
margins_plot(fit, effects = "log_Base4_c:Trt_c", 
             data = mdata, re_formula = NULL, rug = TRUE)
}

}

