% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{me}
\alias{me}
\title{Predictors with measurement error in \pkg{brms} models}
\usage{
me(x, noise = NULL)
}
\arguments{
\item{x}{The variable measured with error.}

\item{noise}{Known measurement error of \code{x}.}
}
\description{
Predictors with measurement error in \pkg{brms} models
}
\details{
This function is almost solely useful when
  called within functions of the \pkg{brms} package.
}

