% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmloo.R
\name{mmloo.brmsfit}
\alias{mmloo.brmsfit}
\alias{mmloo}
\title{Moment matching for efficient approximate leave-one-out cross-validation}
\usage{
\method{mmloo}{brmsfit}(x, loo, ...)
}
\arguments{
\item{x}{An object of class \code{brmsfit}.}

\item{loo}{An object of class \code{loo} originally created from \code{x}.}

\item{...}{Further arguments passed to the underlying methods.}
}
\value{
An object of the class \code{loo}.
}
\description{
Moment matching for efficient approximate leave-one-out cross-validation 
(LOO-CV). See \code{\link[loo:mmloo]{mmloo}} for more details.
}
\details{
The moment matching algorithm requires samples 
  of all variables defined in Stan's \code{parameters} block
  to be saved. Otherwise \code{mmloo} cannot be computed.
  Thus, please set \code{save_all_pars = TRUE} in the call to \code{brm},
  if you are planning to apply \code{mmloo} to your models.
}
\examples{
\dontrun{
fit1 <- brm(count ~ zAge + zBase * Trt + (1|patient),
            data = epilepsy, family = poisson(),
            save_all_pars = TRUE)
# throws warning about some pareto k estimates being too high
(loo1 <- loo(fit1))
(mmloo1 <- mmloo(fit1, loo = loo1, k_thres = 0.7))
}

}
