% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{MultiNormal}
\alias{MultiNormal}
\alias{dmulti_normal}
\alias{rmulti_normal}
\title{The Multivariate Normal Distribution}
\usage{
dmulti_normal(x, mu, Sigma, log = FALSE, check = FALSE)

rmulti_normal(n, mu, Sigma, check = FALSE)
}
\arguments{
\item{x}{Vector or matrix of quantiles. If \code{x} is a matrix,
each row is taken to be a quantile.}

\item{mu}{Mean vector with length equal to the number of dimensions.}

\item{Sigma}{Covariance matrix.}

\item{log}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{check}{Logical; Indicates whether several input checks
should be performed. Defaults to \code{FALSE} to improve
efficiency.}

\item{n}{Number of draws to sample from the distribution.}
}
\description{
Density function and random generation for the multivariate normal
distribution with mean vector \code{mu} and covariance matrix \code{Sigma}.
}
\details{
See the Stan user's manual \url{https://mc-stan.org/documentation/}
for details on the parameterization
}
