% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{MultiStudentT}
\alias{MultiStudentT}
\alias{dmulti_student_t}
\alias{rmulti_student_t}
\title{The Multivariate Student-t Distribution}
\usage{
dmulti_student_t(x, df, mu, Sigma, log = FALSE, check = FALSE)

rmulti_student_t(n, df, mu, Sigma, check = FALSE)
}
\arguments{
\item{x}{Vector or matrix of quantiles. If \code{x} is a matrix,
each row is taken to be a quantile.}

\item{df}{Vector of degrees of freedom.}

\item{mu}{Location vector with length equal to the number of dimensions.}

\item{Sigma}{Covariance matrix.}

\item{log}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{check}{Logical; Indicates whether several input checks
should be performed. Defaults to \code{FALSE} to improve
efficiency.}

\item{n}{Number of draws to sample from the distribution.}
}
\description{
Density function and random generation for the multivariate Student-t
distribution with location vector \code{mu}, covariance matrix \code{Sigma},
and degrees of freedom \code{df}.
}
\details{
See the Stan user's manual \url{https://mc-stan.org/docs/}
  for details on the parameterization
}
