% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsformula.R
\name{mvbind}
\alias{mvbind}
\title{Bind response variables in multivariate models}
\usage{
mvbind(...)
}
\arguments{
\item{...}{Same as in \code{\link{cbind}}}
}
\description{
Can be used to specify a multivariate \pkg{brms} model within a single
formula. Outside of \code{\link{brmsformula}}, it just behaves like
\code{\link{cbind}}.
}
\examples{
bf(mvbind(y1, y2) ~ x)

}
\seealso{
\code{\link{brmsformula}}, \code{\link{mvbrmsformula}}
}
