% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{ngrps.brmsfit}
\alias{ngrps.brmsfit}
\alias{ngrps}
\title{Number of Grouping Factor Levels}
\usage{
\method{ngrps}{brmsfit}(object, ...)

ngrps(object, ...)
}
\arguments{
\item{object}{An \R object.}

\item{...}{Currently ignored.}
}
\value{
A named list containing the number of levels per
  grouping factor.
}
\description{
Extract the number of levels of one or more grouping factors.
}
