% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ngrps}
\alias{ngrps}
\alias{ngrps.brmsfit}
\title{Number of levels}
\usage{
ngrps(object, ...)
}
\arguments{
\item{object}{An \code{R} object typically of class \code{brmsfit}.}

\item{...}{Currently ignored.}
}
\value{
Number(s) of levels
}
\description{
Number of levels of one or more grouping factor
}
\details{
Currently there are methods for \code{brmsfit} objects.
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}

