% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{nsamples.brmsfit}
\alias{nsamples.brmsfit}
\alias{nsamples}
\title{Number of Posterior Samples}
\usage{
\method{nsamples}{brmsfit}(object, subset = NULL, incl_warmup = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}.}

\item{subset}{An optional integer vector defining a subset of samples
to be considered.}

\item{incl_warmup}{A flag indicating whether to also count warmup / burn-in
samples.}

\item{...}{Currently ignored.}
}
\description{
Extract the number of posterior samples stored in a fitted Bayesian model.
}
