% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/generics.R
\name{nsamples.brmsfit}
\alias{nsamples.brmsfit}
\alias{nsamples}
\title{Number of Posterior Samples}
\usage{
\method{nsamples}{brmsfit}(x, subset = NULL, incl_warmup = FALSE, ...)

nsamples(x, ...)
}
\arguments{
\item{x}{An \R object}

\item{subset}{An optional integer vector defining a 
subset of samples to be considered.}

\item{incl_warmup}{A flag indicating whether to also 
count warmup / burn-in samples.}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Extract the number of posterior samples 
stored in a fitted Bayesian model.
}
\details{
Currently there are methods for \code{brmsfit} objects.
}
