% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.generics.R
\name{par.names}
\alias{par.names}
\title{Extract parameter names}
\usage{
par.names(x, ...)
}
\arguments{
\item{x}{An \code{R} object typically of class \code{brmsfit}}

\item{...}{Currently ignored}
}
\value{
A character vector containing the parameter names of the model
}
\description{
Extract all parameter names of a given model
}
\details{
Currently there are methods for \code{brmsfit} objects.
}
\author{
Paul-Christian Buerkner <\email{paul.buerkner@gmail.com}>
}

