% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naming.R
\name{parnames.formula}
\alias{parnames.formula}
\title{Extract parameter names}
\usage{
\method{parnames}{formula}(x, data = NULL, family = "gaussian",
  addition = NULL, autocor = NULL, partial = NULL,
  threshold = "flexible", internal = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{formula}}

\item{data}{An optional data frame, list or environment  (or object coercible by \code{as.data.frame} to a data frame) containing
the variables in the model. If not found in data, the variables are taken from \code{environment(formula)},
typically the environment from which \code{brm} is called. Although it is optional, we strongly recommend to supply a data.frame.}

\item{family}{A vector of one or two character strings. The first string indicates the distribution of the dependent variable (the 'family'). Currently, the following families are supported:
\code{"gaussian"}, \code{"student"}, \code{"cauchy"}, \code{"binomial"}, \code{"bernoulli"}, \code{"categorical"}, \code{"poisson"}, \code{"negbinomial"}, \cr
\code{"geometric"}, \code{"gamma"}, \code{"exponential"}, \code{"weibull"}, \code{"cumulative"}, \code{"cratio"}, \code{"sratio"}, and \code{"acat"}.
The second string indicates the link function, which must be supported by the distribution of the dependent variable.
If not specified, default link functions are used. Further information is provided under 'Details'.}

\item{addition}{A named list of one sided formulas each containing additional information on the response variable. The following names are allowed:
\code{se} for specifying standard errors for meta-analysis, \code{weights} to fit weighted regression models,
\code{trials} to specify the number of trials per observation in binomial models, \code{cat} to specify the number of categories in
categorical or ordinal models, and \code{cens} to indicate censoring. Alternatively, the \code{addition} arguments can be incorporated directly into \code{formula}.
See 'Formula Syntax' under 'Details' for further information.}

\item{autocor}{An optional \code{\link{cor_brms}} object describing the correlation structure within the response variable (i.e. the 'autocorrelation').
See the documentation of \code{\link{cor_brms}} for a description of the available correlation structures.
Defaults to NULL, corresponding to no correlations.}

\item{partial}{A one sided formula of the form \code{~partial.effects} specifing the predictors that can vary between categories in non-cumulative ordinal models
(i.e. in families \code{"cratio"}, \code{"sratio"}, or \code{"acat"}).}

\item{threshold}{A character string indicating the type of thresholds (i.e. intercepts) used in an ordinal model.
\code{"flexible"} provides the standard unstructured thresholds and \code{"equidistant"} restricts the distance between consecutive thresholds to the same value.}

\item{internal}{A flag indicating if the names of additional internal parameters should be displayed.
Setting priors on these parameters is not recommended}

\item{...}{Currently ignored}
}
\value{
A list of character vectors containing the parameter names for which priors may be specified
}
\description{
Extract all parameter names for which priors may be specified
}
\examples{
parnames(rating ~ treat + period + carry + (1+carry|subject),
          data = inhaler, family = "student")

parnames(count ~ log_Age_c + log_Base4_c * Trt_c + (1|patient) + (1|visit),
          data = epilepsy, family = "poisson")
}

