% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pbeta_binom}
\alias{pbeta_binom}
\title{PMF of beta-binomial distribution}
\usage{
pbeta_binom(q, size, mu = 1, phi = 1, lower.tail = TRUE, log.p = FALSE)
}
\description{
PMF of beta-binomial distribution
}
