% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{post_prob.brmsfit}
\alias{post_prob.brmsfit}
\alias{post_prob}
\title{Posterior Model Probabilities from Marginal Likelihoods}
\usage{
\method{post_prob}{brmsfit}(x, ..., prior_prob = NULL, model_names = NULL,
  bs_args = list())
}
\arguments{
\item{x}{A \code{brmsfit} object.}

\item{...}{More \code{brmsfit} objects.}

\item{prior_prob}{Numeric vector with prior model probabilities. 
If omitted, a uniform prior is used (i.e., all models are equally 
likely a priori). The default \code{NULL} corresponds to equal 
prior model weights.}

\item{model_names}{If \code{NULL} (the default) will use model names 
derived from deparsing the call. Otherwise will use the passed 
values as model names.}

\item{bs_args}{A list of additional arguments passed to 
\code{\link[brms:bridge_sampler]{bridge_sampler}}.}
}
\description{
Compute posterior model probabilities from marginal likelihoods.
The \code{brmsfit} method is just a thin wrapper around
the corresponding method for \code{bridge} objects.
}
\details{
More details are provided under
  \code{\link[bridgesampling:post_prob]{post_prob}}.
}
\seealso{
\code{
  \link[brms:bridge_sampler]{bridge_sampler},
  \link[brms:bayes_factor]{bayes_factor}
}
}
