% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/generics.R
\name{posterior_average.brmsfit}
\alias{posterior_average.brmsfit}
\alias{posterior_average}
\title{Posterior samples of parameters averaged across models}
\usage{
\method{posterior_average}{brmsfit}(x, ..., pars = NULL, weights = "loo2",
  nsamples = NULL, model_names = NULL, control = list(),
  as.matrix = FALSE)

posterior_average(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{More fitted model objects or further arguments
passed to the underlying post-processing functions.}

\item{pars}{Names of parameters for which to average across models.
Only those parameters can be averaged that appear in every model.
Defaults to all overlapping parameters.}

\item{weights}{Name of the criterion to compute weights from. 
Should be one of \code{"loo"}, \code{"waic"}, \code{"kfold"}, 
\code{"loo2"} (current default), or \code{"marglik"}. 
For the former three options, Akaike weights will be computed
based on the information criterion values returned by
the respective methods. For \code{"loo2"}, method
\code{\link{loo_model_weights}} will be used to obtain weights. 
For \code{"marglik"}, method \code{\link{post_prob}} 
will be used to compute weights based on log marginal 
likelihood values. Alternatively, \code{weights} can be 
a numeric vector of pre-specified weights.}

\item{nsamples}{Total number of posterior samples to use.}

\item{model_names}{If \code{NULL} (the default) will use model names 
derived from deparsing the call. Otherwise will use the passed 
values as model names.}

\item{control}{Optional \code{list} of further arguments 
passed to the function specified in \code{weights}.}

\item{as.matrix}{Should the output be a \code{matrix} instead of a 
\code{data.frame}? Defaults to \code{FALSE}.}
}
\value{
As \code{data.frame} or \code{matrix} of posterior samples.
}
\description{
Compute posterior samples of parameters averaged across models.
Weighting can be done in various ways, for instance using
Akaike weights based on information criteria or 
marginal likelihoods.
}
\examples{
\dontrun{
# model with 'treat' as predictor
fit1 <- brm(rating ~ treat + period + carry, data = inhaler)
summary(fit1)

# model without 'treat' as predictor
fit2 <- brm(rating ~ period + carry, data = inhaler)
summary(fit2)

# compute model-averaged posteriors of overlapping parameters
posterior_average(fit1, fit2, weights = "waic")
}

}
