% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pp_check}
\alias{pp_check}
\title{Temporary pp_check generic function}
\usage{
pp_check(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments}
}
\description{
Temporary pp_check generic function
}
\details{
See \code{\link[brms:pp_check.brmsfit]{pp_check.brmsfit}} for details.
}

