% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{ppc.brmsfit}
\alias{ppc.brmsfit}
\title{Posterior Predictive Checks for \code{brmsfit} Objects}
\usage{
\method{ppc}{brmsfit}(x, type = "dens_overlay", nsamples = 10,
  subset = NULL, ...)
}
\description{
Posterior Predictive Checks for \code{brmsfit} Objects
}

