% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ppc}
\alias{ppc}
\title{temporary ppc generic function}
\usage{
ppc(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{further arguments}
}
\description{
will be removed as soon as the \pkg{ppcheck} package
has its own generic function
}

