% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.methods.R
\name{print.brmsfit}
\alias{print.brmsfit}
\alias{print.brmssummary}
\title{Print a summary for a fitted model represented by a \code{brmsfit} object}
\usage{
\method{print}{brmsfit}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{brmsfit}}

\item{digits}{The number of significant digits for printing out the summary; defaults to 2.
The effective sample size is always rounded to integers.}

\item{...}{Additional arguments that would be passed to method \code{summary} of \code{brmsfit}.}
}
\description{
Print basic information regarding the fitted model and a summary for the fixed and random effects
estimated by the samples included in a \code{brmsfit} object.
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}

