% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.brmsprior}
\alias{print.brmsprior}
\title{Print method for \code{brmsprior} objects}
\usage{
\method{print}{brmsprior}(x, show_df = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{brmsprior}.}

\item{show_df}{Logical; Print priors as a single
\code{data.frame} (\code{TRUE}) or as a sequence of
sampling statements (\code{FALSE})?}

\item{...}{Currently ignored.}
}
\description{
Print method for \code{brmsprior} objects
}
