% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-helpers.R
\name{print_dims.kfold}
\alias{print_dims.kfold}
\title{Temporary print_dim method for kfold objects}
\usage{
print_dims.kfold(x, ...)
}
\arguments{
\item{x}{a \code{kfold} object}

\item{...}{currently ignored}
}
\description{
Temporary print_dim method for kfold objects
}
