% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.generics.R
\name{prior.samples}
\alias{prior.samples}
\alias{prior.samples.brmsfit}
\title{Extract prior samples}
\usage{
prior.samples(x, parameters = NA, ...)
}
\arguments{
\item{x}{An \code{R} object typically of class \code{brmsfit}}

\item{parameters}{Name of parameters for which posterior samples should be returned, as given by a character vector or regular expressions.
By default, all prior samples are extracted}

\item{...}{Currently ignored}
}
\value{
A data frame containing the prior samples.
}
\description{
Extract prior samples of specified parameters
}
\details{
To make use of this function, the model must contain samples of prior distributions.
 This can be ensured by setting \code{prior.samples = TRUE} in function \code{brm}.
 Currently there are methods for \code{brmsfit} objects.
}
\examples{
\dontrun{
fit_i <- brm(rating ~ treat + period + carry + (1|subject),
             data = inhaler, family = "cumulative",
             prior = list(b = "normal(0,2)"), sample.prior = TRUE)

#extract all prior samples
samples1 <- prior.samples(fit_i)
head(samples1)

#extract prior samples for the fixed effect of \\code{treat}.
samples2 <- posterior.samples(fit_i, "b_treat")
head(samples2)
}
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}

