% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/generics.R
\name{priors.brmsfit}
\alias{priors}
\alias{priors.brmsfit}
\title{Extract priors of a Bayesian model}
\usage{
\method{priors}{brmsfit}(x, all = TRUE, ...)

priors(x, ...)
}
\arguments{
\item{x}{An \R object}

\item{all}{Logical; Show all parameters in the model which may have 
priors (\code{TRUE}) or only those with proper priors (\code{FALSE})?}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
For \code{brmsfit} objects, an object of class \code{brmsprior}.
}
\description{
Extract priors of a Bayesian model
}
\examples{
\dontrun{
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c  
             + (1|patient) + (1|obs), 
           data = epilepsy, family = poisson(), 
           prior = c(prior(student_t(5,0,10), class = b),
                     prior(cauchy(0,2), class = sd)))
                   
priors(fit)
priors(fit, all = FALSE)
print(priors(fit, all = FALSE), show_df = FALSE)
}

}

