% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{read_stanr_fit_as_stanfit}
\alias{read_stanr_fit_as_stanfit}
\title{Read a stanr fit as a brms-formatted stanfit object}
\usage{
read_stanr_fit_as_stanfit(
  fit,
  model = NULL,
  exclude = "",
  algorithm = "sampling"
)
}
\arguments{
\item{fit}{A stanr fit object (\code{StanMCMC}, \code{StanVB},
\code{StanPathfinder}, or \code{StanLaplace}).}

\item{model}{The stanr \code{StanModel} object the fit was generated from
(optional). Provide this argument if you want to allow updating the model
without recompilation.}

\item{exclude}{Character vector of variables to exclude from the stanfit.}

\item{algorithm}{The algorithm with which the model was fitted.
See \code{\link{brm}} for details.}
}
\value{
A stanfit object consistent with the structure of the \code{fit}
 slot of a brmsfit object.
}
\description{
\code{read_stanr_fit_as_stanfit} is used internally to convert the in-memory
fit object returned by a \pkg{stanr} model-fitting method (\code{$sample()},
\code{$variational()}, \code{$pathfinder()}, or \code{$laplace()}) into a
\code{stanfit} object that is consistent with the structure of the fit slot of
a brmsfit object.
}
