% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{recompile_model}
\alias{recompile_model}
\title{Recompile Stan models in \code{brmsfit} objects}
\usage{
recompile_model(x, recompile = NULL)
}
\arguments{
\item{x}{An object of class \code{brmsfit}.}

\item{recompile}{Logical, indicating whether the Stan model should be
recompiled. If \code{NULL} (the default), \code{recompile_model} tries
to figure out internally, if recompilation is necessary. Setting it to
\code{FALSE} will cause \code{recompile_model} to always return the
\code{brmsfit} object unchanged.}
}
\value{
A (possibly updated) \code{brmsfit} object.
}
\description{
Recompile the Stan model inside a \code{brmsfit} object, if necessary.
This does not change the model, it simply recreates the executable
so that sampling is possible again.
}
