% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.methods.R
\name{residuals.brmsfit}
\alias{residuals.brmsfit}
\title{Extract Model Residuals from brmsfit Objects}
\usage{
\method{residuals}{brmsfit}(object, summary = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{summary}{logical. Should summary statistics (i.e. means, sds, and 95\% intervals) be returned
instead of the raw values. Default is \code{FALSE}}

\item{...}{Currently ignored}
}
\value{
Residuals extracted from \code{object}.
}
\description{
Extract Model Residuals from brmsfit Objects
}
\details{
Currently, the method only supports families \code{gaussian}, \code{student},
  or \code{cauchy}. This will likely be changed in the future.
}
\examples{
\dontrun{
## fit the model
fit <- brm(rating ~ treat + period + carry + (1|subject), data = inhaler,
           n.cluster = 2)

## extract residuals
res <- residuals(fit, summary = TRUE)
head(res)
}
}

