% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-ad.R
\name{resp_index}
\alias{resp_index}
\title{(Deprecated) Additional response information for handling response indexes}
\usage{
resp_index(x)
}
\arguments{
\item{x}{A vector of index values that assigns each observation to an
index, which can then be used by other model terms.}
}
\value{
A list of additional response information to be processed further
  by \pkg{brms}.
}
\description{
Provide additional information for handling response indexes
in \pkg{brms} models. This function is deprecated and replaced
by the \code{idx} argument of \code{\link{resp_mi}}.
}
\details{
This function is almost solely useful when called in formulas passed to the
  \pkg{brms} package. Within formulas, the \code{resp_} prefix may be
  omitted. See the Details section of \code{\link{brmsformula}} (under
  "Additional response information") for more information.
}
\seealso{
\code{\link{resp_mi}}, \code{\link{brmsformula}}
}
