% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-ad.R
\name{resp_mi}
\alias{resp_mi}
\title{Additional response information for handling missing values}
\usage{
resp_mi(sdy = NA, idx = NA)
}
\arguments{
\item{sdy}{Optional known measurement error of the response
treated as standard deviation. If specified, handles
measurement error and (completely) missing values
at the same time using the plausible-values-technique.}

\item{idx}{Optional index variable assigning observations with missing responses
to individual latent response values. This allows to express that multiple
observations share the same latent missing value. It is also needed when
using the \code{idx} argument in \code{\link{mi}} predictor terms.}
}
\value{
A list of additional response information to be processed further
  by \pkg{brms}.
}
\description{
Provide additional information on missing values in response variables
in \pkg{brms} models. Required when using \code{\link{mi}} predictor terms.
}
\details{
This function is almost solely useful when called in formulas passed to the
  \pkg{brms} package. Within formulas, the \code{resp_} prefix may be
  omitted. See the Details section of \code{\link{brmsformula}} (under
  "Additional response information") for more information.
  Please also see \code{\link{mi}} for more details and examples.

  It is highly recommended to use single data variables as inputs
  for \code{sdy} and \code{idx} (instead of a more complicated expression)
  to make sure all post-processing functions work as expected.
}
\examples{
\dontrun{
data("nhanes", package = "mice")
N <- nrow(nhanes)

# simple model with missing data
bform1 <- bf(bmi | mi() ~ age * mi(chl)) +
  bf(chl | mi() ~ age) +
  set_rescor(FALSE)

fit1 <- brm(bform1, data = nhanes)

summary(fit1)
plot(conditional_effects(fit1, resp = "bmi"), ask = FALSE)
loo(fit1, newdata = na.omit(fit1$data))
}

}
\seealso{
\code{\link{brmsformula}}, \code{\link{mi}}
}
