% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-ad.R
\name{resp_subset}
\alias{resp_subset}
\title{Additional response information for subsetting data}
\usage{
resp_subset(x)
}
\arguments{
\item{x}{A logical vector or one that can be transformed to it. Observations
for which \code{x} is \code{FALSE} will be excluded from the data used in
that specific univariate model.}
}
\value{
A list of additional response information to be processed further
  by \pkg{brms}.
}
\description{
Provide additional information for subsetting data
in \pkg{brms} models. This is relevant only in multivariate models
where \code{resp_subset} allows to use different subsets of the data
in different univariate models.
}
\details{
This function is almost solely useful when called in formulas passed to the
  \pkg{brms} package. Within formulas, the \code{resp_} prefix may be
  omitted. See the Details section of \code{\link{brmsformula}} (under
  "Additional response information") for more information.
}
\examples{
\dontrun{
# generate a simple dataset
dat <- data.frame(y = rnorm(10), z = rnorm(10),
                  sub = c(rep(TRUE, 5), rep(FALSE, 5)))

# only use the first 5 observations to predict z
bform <- bf(y ~ 1) + bf(z | subset(sub) ~ y) + set_rescor(FALSE)
fit <- brm(bform, dat)
summary(fit)
}

}
