% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure.R
\name{restructure}
\alias{restructure}
\title{Retructure Old \code{brmsfit} Objects}
\usage{
restructure(x, rstr_summary = FALSE)
}
\arguments{
\item{x}{An object of class \code{brmsfit}.}

\item{rstr_summary}{Logical; If \code{TRUE}, the cached summary
stored by \pkg{rstan} is restructured as well.}
}
\value{
A \code{brmsfit} object compatible with the latest version
  of \pkg{brms}.
}
\description{
Restructure old \code{brmsfit} objects to work with 
the latest \pkg{brms} version. This function is called
internally when applying post-processing methods.
However, in order to avoid uncessary run time caused
by the restructuring, I recommend explicitely calling
\code{restructure} once per model after updating \pkg{brms}.
}
