% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{set_inits}
\alias{set_inits}
\title{Init definitions for **brms** models}
\usage{
set_inits(
  distribution,
  class = "b",
  coef = "",
  group = "",
  dpar = "",
  nlpar = ""
)
}
\arguments{
\item{class}{The parameter class. Defaults to \code{"b"}
(i.e. population-level effects).
See 'Details' for other valid parameter classes.}

\item{coef}{Name of the coefficient within the parameter class.}

\item{group}{Grouping factor for group-level parameters.}

\item{dpar}{Name of a distributional parameter.
Only used in distributional models.}

\item{nlpar}{Name of a non-linear parameter.
Only used in non-linear models.}
}
\value{
An object of class `brmsinits` to be used in the `init` argument of [brm]
}
\description{
Define how initial values for specific parameters are generated.
}
\examples{
\dontrun{
inits <- set_inits("normal(0, 1)", class = "Intercept", coef = "mu") +
        set_inits("uniform(-1, 1)", class = "b", coef = "mu")
# use the inits in a brm call
fit <- brm(count ~ Trt + zAge, epilepsy, poisson(), init = inits)
}
}
