% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{SkewNormal}
\alias{SkewNormal}
\alias{dskew_normal}
\alias{pskew_normal}
\alias{rskew_normal}
\title{The Skew-Normal Distribution}
\usage{
dskew_normal(x, mu = 0, sigma = 1, alpha = 0, log = FALSE)

pskew_normal(q, mu = 0, sigma = 1, alpha = 0, log.p = FALSE)

rskew_normal(n, mu = 0, sigma = 1, alpha = 0)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{mu}{Vector of mean values.}

\item{sigma}{Vector of standard deviation values.}

\item{alpha}{Vector of skewness values.}

\item{log}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{log.p}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{n}{Number of samples to draw from the distribution.}
}
\description{
Density, distribution function, and random generation for the 
skew-normal distribution with mean \code{mu}, 
standard deviation \code{sigma}, and skewness \code{alpha}.
}
\details{
See \code{vignette("brms_families")} for details
on the parameterization.
}
