% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{stancode.brmsfit}
\alias{stancode.brmsfit}
\alias{stancode}
\title{Extract Stan model code}
\usage{
\method{stancode}{brmsfit}(object, version = TRUE, ...)

stancode(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{version}{Logical; indicates if the first line containing
the \pkg{brms} version number should be included.
Defaults to \code{TRUE}.}

\item{...}{Currently ignored}
}
\value{
Stan model code for further processing.
}
\description{
Extract Stan model code
}
