% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{stancode}
\alias{stancode}
\title{Extract Stan Model Code}
\usage{
stancode(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{...}{Currently ignored}
}
\value{
model code in stan language for further processing.
}
\description{
Extract the model code in Stan language
}

