% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{standata}
\alias{standata}
\alias{standata.brmsfit}
\title{Extract Data passed to Stan}
\usage{
standata(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{...}{Currently ignored}
}
\value{
A named list containing the data passed to Stan
}
\description{
Extract all data that was used by Stan to fit the model
}

