% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{stanthreads}
\alias{stanthreads}
\title{Threading in Stan}
\usage{
stanthreads(threads = NULL, grainsize = NULL, static = FALSE)
}
\arguments{
\item{threads}{TODO}

\item{grainsize}{TODO}

\item{static}{Logical. TODO}
}
\value{
A \code{stanthreads} object which can be passed to the
  \code{threads} argument of \code{brms} and related functions.
}
\description{
Use threads for within-chain parallelization in Stan.
}
