% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Student}
\alias{Student}
\alias{dstudent}
\alias{pstudent}
\alias{qstudent}
\alias{rstudent}
\title{The Student-t Distribution}
\usage{
dstudent(x, df, mu = 0, sigma = 1, log = FALSE)

pstudent(q, df, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qstudent(p, df, mu = 0, sigma = 1)

rstudent(n, df, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{df}{Vector of degrees of freedom.}

\item{mu}{Vector of location values.}

\item{sigma}{Vector of scale values.}

\item{log, log.p}{Logical; If \code{TRUE} values are returned on the log scale.}

\item{lower.tail}{Logical; If \code{TRUE} (default), probabilities are 
P[X ≤ x], otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length 
is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation 
for the Student-t distribution with location \code{mu}, scale \code{sigma},
and degrees of freedom \code{df}.
}
\details{
See \code{vignette("brms_families")} for details
on the parameterization.
}
\seealso{
\code{\link[stats:TDist]{TDist}}
}
