% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_default}
\alias{theme_default}
\title{Default \pkg{bayesplot} Theme for \pkg{ggplot2} Graphics}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
A \code{theme} object used in \pkg{ggplot2} graphics.
}
\description{
This theme is imported from the \pkg{bayesplot} package.
See \code{\link[bayesplot:theme_default]{theme_default}}
for a complete documentation.
}
