% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{threading}
\alias{threading}
\title{Threading in Stan}
\usage{
threading(threads = NULL, grainsize = NULL, static = FALSE)
}
\arguments{
\item{threads}{TODO}

\item{grainsize}{TODO}

\item{static}{Logical. TODO}
}
\value{
A \code{brmsthreads} object which can be passed to the
  \code{threads} argument of \code{brm} and related functions.
}
\description{
Use threads for within-chain parallelization in \pkg{Stan} via the \pkg{brms}
interface.
}
