% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ZeroOneInflated}
\alias{ZeroOneInflated}
\alias{dzero_one_inflated_beta}
\alias{pzero_one_inflated_beta}
\alias{qzero_one_inflated_beta}
\title{Zero-One-Inflated Beta Distribution}
\usage{
dzero_one_inflated_beta(x, shape1, shape2, zoi, coi, log = FALSE)

pzero_one_inflated_beta(
  q,
  shape1,
  shape2,
  zoi,
  coi,
  lower.tail = TRUE,
  log.p = FALSE
)

qzero_one_inflated_beta(
  p,
  shape1,
  shape2,
  zoi,
  coi,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{shape1, shape2}{shape parameters of the beta component}

\item{zoi}{zero-one-inflation probability}

\item{coi}{conditional one-inflation probability}

\item{log}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{q}{Vector of quantiles.}

\item{lower.tail}{Logical; If \code{TRUE} (default), return P(X <= x).
Else, return P(X > x) .}

\item{log.p}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{p}{Vector of probabilities.}
}
\description{
Density, distribution function and quantile function for the
zero-one-inflated beta distribution.
}
\details{
With probability \eqn{\zeta}{zoi} the response is 0 or 1, and conditionally
on inflation it is 1 with probability \eqn{\gamma}{coi}. With probability
\eqn{1 - \zeta}{1 - zoi} the response follows a beta distribution.
}
