% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_mixture.R
\name{zi_families}
\alias{zi_families}
\title{List of all true zero-inflated families, where observed zeros may arise
either from the zero-inflation component or from the other component.}
\usage{
zi_families()
}
\description{
List of all true zero-inflated families, where observed zeros may arise
either from the zero-inflation component or from the other component.
}
