% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/generics.R
\name{bayes_factor.brmsfit}
\alias{bayes_factor.brmsfit}
\alias{bayes_factor}
\title{Bayes Factor(s) for Marginal Likelihoods}
\usage{
\method{bayes_factor}{brmsfit}(x1, x2, log = FALSE, ...)

bayes_factor(x1, x2, ...)
}
\arguments{
\item{x1}{An \R object}

\item{x2}{Another \R object}

\item{log}{Report Bayes factors on the log-scale?}

\item{...}{More arguments passed to or from other methods.}
}
\description{
Generic function that computes Bayes factor(s) from marginal likelihoods.
}
