% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.old.R
\name{brm.plot}
\alias{brm.plot}
\title{Trace and density plots for MCMC samples (deprecated)}
\usage{
brm.plot(fit, family = "^[^(r_)]")
}
\arguments{
\item{fit}{An object of class \code{stanfit} or an object that can be coerced to an
\code{mcmc.list} via \code{as.mcmc} of package \code{coda}.}

\item{family}{Name of the family of parameters to plot, as given by a character vector or a regular expression. A family of parameters is considered to
  be any group of parameters with the same name but different numerical value between square brackets (as beta[1], beta[2], etc).
  By default, all parameters except for random effects of each level are plotted.
  For details see the documentation of \code{ggmcmc}.}
}
\description{
Trace and density plots for MCMC samples (deprecated)
}
\details{
This function is deprecated and will soon be defunct. For models created with brms 0.2.0 or higher,
we recommend using the S3 plot method for class \code{brmsfit} (see also \code{\link[brms:plot.brmsfit]{plot.brmsfit}}).
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}

