% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_multiple.R
\name{brm_multiple}
\alias{brm_multiple}
\title{Run the same \pkg{brms} model on multiple datasets}
\usage{
brm_multiple(formula, data, family = gaussian(), prior = NULL,
  autocor = NULL, cov_ranef = NULL, sample_prior = c("no", "yes", "only"),
  sparse = FALSE, knots = NULL, stanvars = NULL, stan_funs = NULL,
  combine = TRUE, seed = NA, ...)
}
\arguments{
\item{formula}{An object of class 
\code{\link[stats:formula]{formula}},
\code{\link{brmsformula}}, or \code{\link{mvbrmsformula}}
(or one that can be coerced to that classes): 
A symbolic description of the model to be fitted. 
The details of model specification are explained in 
\code{\link{brmsformula}}.}

\item{data}{A list of data.frames each of which will
be used to fit a separate model. Alternatively, a 
\code{mids} object from the \pkg{mice} package.}

\item{family}{A description of the response distribution and link function 
to be used in the model. This can be a family function, 
a call to a family function or a character string naming the family.
Every family function has a \code{link} argument allowing to specify
the link function to be applied on the response variable.
If not specified, default links are used.
For details of supported families see 
\code{\link{brmsfamily}}.
By default, a linear \code{gaussian} model is applied.
In multivariate models, \code{family} might also be a list of families.}

\item{prior}{One or more \code{brmsprior} objects created by
\code{\link{set_prior}} or related functions 
and combined using the \code{c} method or the \code{+} operator.
See also  \code{\link{get_prior}} for more help.}

\item{autocor}{An optional \code{\link{cor_brms}} object describing 
the correlation structure within the response variable 
(i.e., the 'autocorrelation'). 
See the documentation of \code{\link{cor_brms}} for a description 
of the available correlation structures. Defaults to \code{NULL}, 
corresponding to no correlations.
In multivariate models, \code{autocor} might also be a list 
of autocorrelation structures.}

\item{cov_ranef}{A list of matrices that are proportional to the 
(within) covariance structure of the group-level effects. 
The names of the matrices should correspond to columns 
in \code{data} that are used as grouping factors. 
All levels of the grouping factor should appear as rownames 
of the corresponding matrix. This argument can be used,
among others to model pedigrees and phylogenetic effects.
See \code{vignette("brms_phylogenetics")} for more details.}

\item{sample_prior}{Indicate if samples from all specified 
proper priors should be drawn additionally to the posterior samples
(defaults to \code{"no"}). Among others, these samples can be used 
to calculate Bayes factors for point hypotheses. 
If set to \code{"only"}, samples are drawn solely from
the priors ignoring the likelihood. In this case, 
all parameters must have proper priors.}

\item{sparse}{Logical; indicates whether the population-level 
design matrices should be treated as sparse (defaults to \code{FALSE}). 
For design matrices with many zeros, this can considerably 
reduce required memory. Sampling speed is currently not 
improved or even slightly decreased.}

\item{knots}{Optional list containing user specified knot values to be 
used for basis construction of smoothing terms. 
See \code{\link[mgcv:gamm]{gamm}} for more details.}

\item{stanvars}{An optional \code{stanvars} object generated
by function \code{\link{stanvar}} to define additional variables
for use in \pkg{Stan}'s program blocks.}

\item{stan_funs}{(Deprecated) An optional character string containing 
self-defined  \pkg{Stan} functions, which will be included in the functions 
block of the generated \pkg{Stan} code. It is now recommended to use 
the \code{stanvars} argument for this purpose, instead.}

\item{combine}{Logical; Indicates if the fitted models
should be combined into a single fitted model object
via \code{\link{combine_models}}. Defaults to \code{TRUE}.}

\item{seed}{The seed for random number generation to make results
reproducible. If \code{NA} (the default), \pkg{Stan} will set
the seed randomly.}

\item{...}{Further arguments passed to \code{\link{brm}}.}
}
\value{
If \code{combine = TRUE} a \code{brmsfit_multiple} object,
which inherits from class \code{brmsfit} and behaves essentially 
the same. If \code{combine = FALSE} a list of \code{brmsfit} objects.
}
\description{
Run the same \pkg{brms} model on multiple datasets and
then combine the results into one fitted model object. 
This is useful in particular for multiple missing value imputation,
where the same model is fitted on multiple imputed data sets.
}
\details{
The combined model may issue false positive 
convergence warnings, as the MCMC chains corresponding
to different datasets may not necessarily overlap, even 
if each of the original models did converge.
To find out whether each of the original models converged,
investigate \code{fit$rhats}, where \code{fit} 
denotes the output of \code{brm_multiple}.
}
\examples{
\dontrun{
library(mice)
imp <- mice(nhanes2)

# fit the model using mice and lm
fit_imp1 <- with(lm(bmi~age+hyp+chl), data = imp)
summary(pool(fit_imp1))

# fit the model using brms
fit_imp2 <- brm_multiple(bmi~age+hyp+chl, data = imp, chains = 1)
summary(fit_imp2)
plot(fit_imp2, pars = "^b_")
# investigate convergence of the original models
fit_imp2$rhats
}

}
