% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/families.R
\name{brmsfamily}
\alias{acat}
\alias{bernoulli}
\alias{brmsfamily}
\alias{categorical}
\alias{cauchy}
\alias{cratio}
\alias{cumulative}
\alias{exponential}
\alias{geometric}
\alias{negbinomial}
\alias{sratio}
\alias{student}
\alias{weibull}
\title{Special Family Functions for \pkg{brms} Models}
\usage{
student(link = "identity")

cauchy(link = "identity")

bernoulli(link = "logit")

negbinomial(link = "log")

geometric(link = "log")

exponential(link = "log")

weibull(link = "log")

categorical(link = "logit")

cumulative(link = "logit")

sratio(link = "logit")

cratio(link = "logit")

acat(link = "logit")
}
\arguments{
\item{link}{A specification for the model link function. This can be a name/expression,
  a literal character string or a length-one character vector. For a full list of link functions
  supported by families applied \pkg{brms}, see the 'Details' section of \code{\link[brms:brm]{brm}}.}
}
\description{
Family objects provide a convenient way to specify the details of the models
used by many model fitting functions. The families present here are currently
for use with \pkg{brms} only and will NOT work with other model fitting functions such as
\code{glm} or \code{glmer}. However, the standard family functions as decribed in
\code{\link[stats:family]{family}} will work with \pkg{brms}. For a full list
of families and link functions supported by \pkg{brms}, see the documentation
(in particular the 'Details' section) of \code{\link[brms:brm]{brm}}.
}

