% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{coef.brmsfit}
\alias{coef.brmsfit}
\title{Extract Model Coefficients}
\usage{
\method{coef}{brmsfit}(object, summary = TRUE, robust = FALSE,
  probs = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{summary}{Should summary statistics 
 (i.e. means, sds, and 95\% intervals) be returned
instead of the raw values? Default is \code{TRUE}.}

\item{robust}{If \code{FALSE} (the default) the mean is used as 
the measure of central tendency and the standard deviation as 
the measure of variability. If \code{TRUE}, the median and the 
median absolute deivation (MAD) are applied instead.
Only used if \code{summary} is \code{TRUE}.}

\item{probs}{The percentiles to be computed by the \code{quantile} 
function. Only used if \code{summary} is \code{TRUE}.}

\item{...}{Currently ignored.}
}
\value{
If \code{old} is \code{FALSE}: A list of arrays 
 (one per grouping factor). If \code{summary} is \code{TRUE},
 names of the first dimension are the factor levels and names
 of the third dimension are the group-level effects. 
 If \code{summary} is \code{FALSE}, names of the second dimension
 are the factor levels and names of the third dimension are the 
 group-level effects.
}
\description{
Extract model coefficients, which are the sum of population-level 
effects and corresponding group-level effects
}
\examples{
\dontrun{
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c + (1+Trt_c|visit), 
           data = epilepsy, family = gaussian(), chains = 2)
## extract population and group-level coefficients separately
fixef(fit)
ranef(fit)
## extract combined coefficients 
coef(fit)
}

}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}
