% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlations.R
\name{cor_ar}
\alias{cor.ar}
\alias{cor_ar}
\title{AR(p) correlation structure}
\usage{
cor_ar(formula = ~1, p = 1)
}
\arguments{
\item{formula}{A one sided formula of the form ~ t, or ~ t | g, specifying a time covariate t and, optionally, a grouping factor g.
A covariate for this correlation structure must be integer valued. When a grouping factor is present in \code{formula}, the correlation structure
is assumed to apply only to observations within the same grouping level; observations with different grouping levels are assumed to be uncorrelated.
Defaults to ~ 1, which corresponds to using the order of the observations in the data as a covariate, and no groups.}

\item{p}{A non-negative integer specifying the autoregressive order of the ARMA structure. Default is 1.}
}
\value{
An object of class \code{cor_arma} containing solely autoregression terms.
}
\description{
This function is a constructor for the \code{cor_arma} class, allowing for autoregression terms only.
}
\details{
The AR structure implemented in \pkg{brms} regresses previous response(s) on the current response.
  Other packages may use the term AR for the regression of previous residual(s) on the current residual.
  When comparing the outputs of different packages using AR coefficients, it is thus important to
  check which AR structure (response or residual) was actually applied.
  AR effects of residuals are currently developed for \pkg{brms}.
}
\examples{
cor_ar(~visit|patient, p = 2)
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}
\seealso{
\code{\link{cor_arma}}
}

