% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{draws-index-brms}
\alias{draws-index-brms}
\alias{variables}
\alias{nvariables}
\alias{niterations}
\alias{nchains}
\alias{ndraws}
\alias{Index}
\alias{variables,}
\alias{iterations,}
\alias{chains,}
\alias{and}
\alias{draws.}
\alias{variables.brmsfit}
\alias{nvariables.brmsfit}
\alias{niterations.brmsfit}
\alias{nchains.brmsfit}
\alias{ndraws.brmsfit}
\title{Index \code{brmsfit} objects}
\usage{
\method{variables}{brmsfit}(x, ...)

\method{nvariables}{brmsfit}(x, ...)

\method{niterations}{brmsfit}(x)

\method{nchains}{brmsfit}(x)

\method{ndraws}{brmsfit}(x)
}
\arguments{
\item{x}{A \code{brmsfit} object or another \R object for which
the methods are defined.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\description{
Index \code{brmsfit} objects
}
