% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{fitted.brmsfit}
\alias{fitted.brmsfit}
\title{Extract Model Fitted Values of \code{brmsfit} Objects}
\usage{
\method{fitted}{brmsfit}(object, summary = TRUE, probs = c(0.025, 0.975),
  ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{summary}{logical. Should summary statistics (i.e. means, sds, and 95\% intervals) be returned
instead of the raw values. Default is \code{FALSE}}

\item{probs}{The percentiles to be computed by the \code{quantile} function. Only used if \code{summary = TRUE}.}

\item{...}{Currently ignored}
}
\value{
Fitted values extracted from \code{object}. If \code{summary = TRUE} this is a S x N matrix and if \code{summary = FALSE}
  a N x C matrix, where S is the number of samples, N is the number of observations,
  and C is equal to \code{length(probs) + 2}.
}
\description{
Extract Model Fitted Values of \code{brmsfit} Objects
}
\details{
Currently, the method does not support \code{categorical} or ordinal models.
}
\examples{
\dontrun{
## fit a model
fit <- brm(rating ~ treat + period + carry + (1|subject), data = inhaler,
           n.cluster = 2)

## extract fitted values
fitted_values <- fitted(fit)
head(fitted_values)
}
}

