% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{fitted.brmsfit}
\alias{fitted.brmsfit}
\title{Extract Model Fitted Values of \code{brmsfit} Objects}
\usage{
\method{fitted}{brmsfit}(object, newdata = NULL, re_formula = NULL,
  scale = c("response", "linear"), allow_new_levels = FALSE,
  summary = TRUE, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{newdata}{An optional data.frame containing new data to make predictions for.
If \code{NULL} (default), the data used to fit the model is applied.}

\item{re_formula}{formula for random effects to condition on.
If \code{NULL} (default), include all random effects; if \code{NA}, include no random effects.
Other options will be implemented in the future.}

\item{scale}{Either \code{"response"} or \code{"linear"}. If \code{scale = "response"} results are returned
on the scale of the response variable. If \code{scale = "linear"} fitted values are returned on the scale of the linear predictor.}

\item{allow_new_levels}{Currenly, \code{FALSE} (no new levels allowed) is the only option. This will change in
future versions of the package.}

\item{summary}{logical. Should summary statistics (i.e. means, sds, and 95\% intervals) be returned
instead of the raw values. Default is \code{TRUE}}

\item{probs}{The percentiles to be computed by the \code{quantile} function. Only used if \code{summary = TRUE}.}

\item{...}{Currently ignored}
}
\value{
Fitted values extracted from \code{object}. The output depends on the family:
  If \code{summary = TRUE} it is a N x E x C array for categorical and ordinal models and a N x E matrix else.
  If \code{summary = FALSE} it is a S x N x C array for categorical and ordinal models and a S x N matrix else.
  N is the number of observations, S is the number of samples, C is the number of categories,
  and E is equal to \code{length(probs) + 2}.
}
\description{
Extract Model Fitted Values of \code{brmsfit} Objects
}
\examples{
\dontrun{
## fit a model
fit <- brm(rating ~ treat + period + carry + (1|subject), data = inhaler,
           n.cluster = 2)

## extract fitted values
fitted_values <- fitted(fit)
head(fitted_values)

## plot fitted means against actual response
dat <- as.data.frame(cbind(Y = fit$data$Y, fitted_values))
ggplot(dat) + geom_point(aes(x = Estimate, y = Y))
}
}

