% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_prior}
\alias{get_prior}
\title{Get default priors for a Bayesian multilevel model}
\usage{
get_prior(formula, data, ...)
}
\arguments{
\item{formula}{A formula object whose class will determine which method will
be used. A symbolic description of the model to be fitted.}

\item{data}{An object of class data.frame, or one that can be coerced to that
class) containing data of all variables used in the model.}

\item{...}{Further arguments passed to the specific method}
}
\value{
A data.frame with columns \code{prior}, \code{class}, \code{coef},
  and \code{group} and several rows, each providing information on a
  parameter (or parameter class) on which priors can be specified. The prior
  column is empty except for internal default priors.
}
\description{
\code{get_prior} is a generic function that can be used to
  obtain the default priors for a Bayesian multilevel model from various
  packages that use the \pkg{brms} package for fitting the model. The
  function invokes particular methods which depend on the class of the
  formula object.

  You can view the available methods by typing \code{methods(get_prior)}.

  See \code{\link[brms:get_prior.default]{get_prior}} for the default method
  applied for \pkg{brms}
}
\examples{
## get all parameters and parameters classes to define priors on
 (prior <- get_prior(count ~ zAge + zBase * Trt + (1|patient) + (1|obs),
                     data = epilepsy, family = poisson()))

## for more examples, see ?get_prior.default for \pkg{brms} and for the other
## methods by first calling:
methods(get_prior)

## and then ?get_prior.* where * is the method name

}
\seealso{
\code{\link{set_prior}} \code{\link{get_prior.default}}
}
