% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-helpers.R
\name{gp}
\alias{gp}
\title{Set up Gaussian process terms in \pkg{brms}}
\usage{
gp(..., cov = "exp_quad")
}
\arguments{
\item{...}{One or more terms containing predictores for
the Gaussian process.}

\item{cov}{Name of the covariance kernel. By default, 
the exponentiated-quadratic kernel \code{"exp_quad"} is used.}
}
\value{
An object of class \code{'gpterm'}, which is a list 
  of arguments to be interpreted by the formula 
  parsing functions of \code{brms}.
}
\description{
Function used to set up a Gaussian process term in \pkg{brms}.
The function does not evaluate its arguments --
it exists purely to help set up a model with Gaussian process terms.
}
\details{
TODO: Explain the idea of Gaussian processes

 Currently, only a single predictor may be passed
 and the exponentiated-quadratic kernel is the only supported 
 covariance kernel. More options will follow in the future.
}
\examples{
\dontrun{
# simulate data using the mgcv package
dat <- mgcv::gamSim(1, n = 30, scale = 2)

# fit a simple gaussian process model
fit1 <- brm(y ~ gp(x2), dat)
summary(fit1)
me1 <- marginal_effects(fit1, nsamples = 200, spaghetti = TRUE)
plot(me1, points = TRUE)

# fit a more complicated gaussian process model
fit2 <- brm(y ~ gp(x0) + x1 + gp(x2) + x3, dat)
summary(fit2)
me2 <- marginal_effects(fit2, nsamples = 200, spaghetti = TRUE)
plot(me2, points = TRUE)

# compare model fit
LOO(fit1, fit2)
}

}
\seealso{
\code{\link[brms:brmsformula]{brmsformula}}
}
