% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm-hash.R
\name{hash_brm_call_master}
\alias{hash_brm_call_master}
\title{Stable hash for a set of brm() arguments}
\usage{
hash_brm_call_master(call, algo = "xxhash64")
}
\arguments{
\item{algo}{Digest algorithm passed to \code{digest}.}

\item{args_list}{A **named** list containing the arguments that uniquely
define a model (e.g., formula, data, family, prior, …).}
}
\value{
A character hash key.
}
\description{
Stable hash for a set of brm() arguments
}
