% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_logit_scaled}
\alias{inv_logit_scaled}
\title{Scaled inverse logit-link}
\usage{
inv_logit_scaled(x, lb = 0, ub = 1)
}
\arguments{
\item{x}{A numeric or complex vector.}

\item{lb}{Lower bound defaulting to \code{0}.}

\item{ub}{Upper bound defaulting to \code{1}.}
}
\value{
A numeric or complex vector between \code{lb} and \code{ub}.
}
\description{
Computes \code{inv_logit(x) * (ub - lb) + lb}
}
