% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{is.cor_brms}
\alias{is.cor_brms}
\alias{is.cor_arma}
\alias{is.cor_sar}
\alias{is.cor_fixed}
\alias{is.cor_bsts}
\title{Check if argument is a correlation structure}
\usage{
is.cor_brms(x)

is.cor_arma(x)

is.cor_sar(x)

is.cor_fixed(x)

is.cor_bsts(x)
}
\arguments{
\item{x}{An \R object.}
}
\description{
Check if argument is one of the correlation structures
used in \pkg{brms}.
}
