% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{lasso}
\alias{lasso}
\title{(Defunct) Set up a lasso prior in \pkg{brms}}
\usage{
lasso(df = 1, scale = 1)
}
\arguments{
\item{df}{Degrees of freedom of the chi-square prior of the inverse tuning
parameter. Defaults to \code{1}.}

\item{scale}{Scale of the lasso prior. Defaults to \code{1}.}
}
\value{
An error indicating that the lasso prior is no longer supported.
}
\description{
This functionality is no longer supported as of brms version 2.19.2. Please
use the \code{\link{horseshoe}} or \code{\link{R2D2}} shrinkage priors instead.
}
\references{
Park, T., & Casella, G. (2008). The Bayesian Lasso. Journal of the American
   Statistical Association, 103(482), 681-686.
}
\seealso{
\code{\link{set_prior}}, \code{\link{horseshoe}}, \code{\link{R2D2}}
}
