% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{logLik.brmsfit}
\alias{logLik.brmsfit}
\title{Compute the pointwise log-likelihood}
\usage{
\method{logLik}{brmsfit}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{brmsfit}.}

\item{...}{Currently ignored}
}
\value{
Usually, an S x N matrix containing the pointwise log-likelihood samples,
        where S is the number of samples and N is the number of observations in the data.
}
\description{
Compute the pointwise log-likelihood
}

