% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{loo_subsample.brmsfit}
\alias{loo_subsample.brmsfit}
\title{Efficient approximate leave-one-out cross-validation (LOO) using subsampling}
\usage{
\method{loo_subsample}{brmsfit}(x, ..., compare = TRUE, resp = NULL,
  reloo = FALSE, k_threshold = 0.7, reloo_args = list(),
  model_names = NULL)
}
\description{
Efficient approximate leave-one-out cross-validation (LOO) using subsampling
}
