% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{make_stancode}
\alias{make_stancode}
\title{Stan Code for \pkg{brms} Models}
\usage{
make_stancode(formula, data = NULL, family = "gaussian",
  prior = prior_frame(), partial = NULL, threshold = "flexible",
  cov.ranef = NULL, sample.prior = FALSE, autocor = cor_arma(),
  save.model = NULL)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.
The details of model specification are given under 'Details'.}

\item{data}{An optional data frame, list or environment  (or object coercible by
\code{as.data.frame} to a data frame) containing the variables in the model. If not found in data,
the variables are taken from \code{environment(formula)},
typically the environment from which \code{brm} is called.
Although it is optional, we strongly recommend to supply a data.frame.}

\item{family}{A description of the error distribution and link function to be used in the model.
This can be a vector of one or two character strings.
The first string indicates the distribution of the dependent variable (the 'family').
Currently, the following families are supported:
\code{"gaussian"}, \code{"student"}, \code{"cauchy"}, \code{"binomial"}, \code{"bernoulli"},
\code{"categorical"}, \code{"poisson"}, \code{"negbinomial"}, \code{"geometric"}, \cr
\code{"gamma"}, \code{"inverse.gaussian"}, \code{"exponential"}, \code{"weibull"},
\code{hurdle_poisson}, \code{hurdle_negbinomial}, \code{hurdle_gamma},
\code{zero_inflated_poisson}, \cr
\code{zero_inflated_negbinomial}, \code{"cumulative"}, \code{"cratio"},
\code{"sratio"}, and \code{"acat"}.
The second string indicates the link function, which must be supported by the family
(if not specified, default links are used).
Alternatively, a family function or the result of a call to a family function are also accepted
(see \code{\link[stats:family]{family}} for help on standard family functions and
\code{\link[brms:brmsfamily]{brmsfamily}} for brms specific family functions).
Further information is provided under 'Details'.}

\item{prior}{One or more \code{brmsprior} objects created by function
\code{\link[brms:set_prior]{set_prior}} and combined using the \code{c} method.
A single \code{brmsprior} object may be passed without \code{c()} surrounding it.
See also  \code{\link[brms:get_prior]{get_prior}} for more help.}

\item{partial}{A one sided formula of the form \code{~expression} specifying the predictors with
category specific effects in non-cumulative ordinal models
(i.e. in families \code{"cratio"}, \code{"sratio"}, or \code{"acat"}).}

\item{threshold}{A character string indicating the type of thresholds (i.e. intercepts)
used in an ordinal model. \code{"flexible"} provides the standard unstructured thresholds and
\code{"equidistant"} restricts the distance between consecutive thresholds to the same value.}

\item{cov.ranef}{A list of matrices that are proportional to the (within) covariance structure of the random effects.
The names of the matrices should correspond to columns in \code{data} that are used as grouping factors.
All levels of the grouping factor should appear as rownames of the corresponding matrix.}

\item{sample.prior}{A flag to indicate if samples from all specified proper priors
should be additionally drawn. Among others, these samples can be used to calculate
Bayes factors for point hypotheses. Default is \code{FALSE}.}

\item{autocor}{An optional \code{\link{cor_brms}} object describing the correlation structure
within the response variable (i.e. the 'autocorrelation').
See the documentation of \code{\link{cor_brms}} for a description
of the available correlation structures. Defaults to NULL, corresponding to no correlations.}

\item{save.model}{Either \code{NULL} or a character string. In the latter case, the model code is
saved in a file named after the string supplied in \code{save.model},
which may also contain the full path where to save the file.
If only a name is given, the file is save in the current working directory.}
}
\value{
A character string containing the fully commented Stan code
  used to fit models with the \code{\link[brms:brm]{brm}} function.
}
\description{
Generate Stan code for \pkg{brms} models
}
\examples{
make_stancode(rating ~ treat + period + carry + (1|subject),
              data = inhaler, family = "cumulative")

make_stancode(count ~ log_Age_c + log_Base4_c * Trt_c
              + (1|patient) + (1|visit),
              data = epilepsy, family = "poisson")
}

