% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stancode.R
\name{make_stancode}
\alias{make_stancode}
\title{Stan Code for a Bayesian multilevel model}
\usage{
make_stancode(formula, data, ...)
}
\arguments{
\item{formula}{A formula object whose class will determine which method will
be used. A symbolic description of the model to be fitted.}

\item{data}{An object of class data.frame, or one that can be coerced to that
class) containing data of all variables used in the model.}

\item{...}{Further arguments passed to the specific method}
}
\value{
A named list of objects containing the required data to fit a
  \pkg{brms} model with \pkg{Stan}.
}
\description{
\code{make_stancode} is a generic function that can be used to
  generate the Stan code to fit Bayesian multilevel models from various packages
  with Stan. The function invokes particular methods which depend on the
  class of the formula object.

  You can view the available methods by typing:

  \code{methods(make_stancode)}

  See \code{\link[brms:make_stancode.default]{make_stancode}} for the default
  method applied for \pkg{brms}
}
\examples{
make_stancode(rating ~ treat + period + carry + (1|subject),
              data = inhaler, family = "cumulative")

## for more examples, see ?make_stancode.default for \pkg{brms} and for the other
## methods by first calling:
methods(make_stancode)

## and then ?make_stancode.* where * is the method name

}
\seealso{
\code{\link{make_stancode.default}}
}
