% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-helpers.R
\name{mi}
\alias{mi}
\title{Predictors with Missing Values in \pkg{brms} Models}
\usage{
mi(x)
}
\arguments{
\item{x}{The variable containing missings.}
}
\description{
Predictors with Missing Values in \pkg{brms} Models
}
\details{
For detailed documentation see \code{help(brmsformula)}. 

This function is almost solely useful when
called in formulas passed to the \pkg{brms} package.
}
\examples{
\dontrun{
# sample some data
N <- 100
dat <- data.frame(y = rnorm(N), x = rnorm(N), z = rnorm(N))
dat$x[sample(1:N, 10)] <- NA
# impute missing values in x and use it as a predictor
bform <- bf(y ~ mi(x)*z) + bf(x | mi() ~ z) + set_rescor(FALSE)
fit <- brm(bform, data = dat)
summary(fit)
} 

}
\seealso{
\code{\link{brmsformula}}
}
