% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-sp.R
\name{mi}
\alias{mi}
\title{Predictors with Missing Values in \pkg{brms} Models}
\usage{
mi(x)
}
\arguments{
\item{x}{The variable containing missings.}
}
\description{
Specify predictor term with missing values in \pkg{brms}. The function does
not evaluate its arguments -- it exists purely to help set up a model.
}
\details{
For detailed documentation see \code{help(brmsformula)}.
}
\examples{
\dontrun{
data("nhanes", package = "mice")
bform <- bf(bmi | mi() ~ age * mi(chl)) +
  bf(chl | mi() ~ age) + set_rescor(FALSE)
fit <- brm(bform, data = nhanes)
summary(fit)
plot(marginal_effects(fit, resp = "bmi"), ask = FALSE)
LOO(fit, newdata = na.omit(fit$data))
} 

}
\seealso{
\code{\link{brmsformula}}
}
