% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-sp.R
\name{mi}
\alias{mi}
\title{Predictors with Missing Values in \pkg{brms} Models}
\usage{
mi(x, idx = NA)
}
\arguments{
\item{x}{The variable containing missing values.}

\item{idx}{An optional variable containing indices of observations in `x`
that are to be used in the model. This is mostly relevant in partially
subsetted models (via \code{\link[addition-terms]{resp_subset}}) but may
also have other applications.}
}
\description{
Specify predictor term with missing values in \pkg{brms}. The function does
not evaluate its arguments -- it exists purely to help set up a model.
}
\details{
For detailed documentation see \code{help(brmsformula)}.
}
\examples{
\dontrun{
data("nhanes", package = "mice")

bform <- bf(bmi | mi() ~ age * mi(chl)) +
  bf(chl | mi() ~ age) + set_rescor(FALSE)
  
fit <- brm(bform, data = nhanes)

summary(fit)
plot(conditional_effects(fit, resp = "bmi"), ask = FALSE)
loo(fit, newdata = na.omit(fit$data))
} 

}
\seealso{
\code{\link{brmsformula}}
}
