% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{mixture}
\alias{mixture}
\title{Finite Mixture Families in \pkg{brms}}
\usage{
mixture(..., flist = NULL, nmix = 1, order = NULL, refcat = NULL, gr = NULL)
}
\arguments{
\item{...}{One or more objects providing a description of the
response distributions to be combined in the mixture model.
These can be family functions, calls to family functions or
character strings naming the families. For details of supported
families see \code{\link{brmsfamily}}.}

\item{flist}{Optional list of objects, which are treated in the
same way as objects passed via the \code{...} argument.}

\item{nmix}{Optional numeric vector specifying the number of times
each family is repeated. If specified, it must have the same length
as the number of families passed via \code{...} and \code{flist}.}

\item{order}{Ordering constraint to identify mixture components.
If \code{'mu'} or \code{TRUE}, population-level intercepts
of the mean parameters are ordered in non-ordinal models
and fixed to the same value in ordinal models (see details).
If \code{'none'} or \code{FALSE}, no ordering constraint is applied.
If \code{NULL} (the default), \code{order} is set to \code{'mu'}
if all families are the same and \code{'none'} otherwise.
Other ordering constraints may be implemented in the future.}

\item{refcat}{Optional reference category for the mixing proportions.
By default (\code{NULL}), when the mixing proportions are predicted,
all but one of them have to be modeled while the remaining one serves
as the reference category (its linear predictor is fixed to zero) to
identify the model. If \code{refcat = NA}, no reference category is used
and all mixing proportions are predicted via a 'softmax' transformation.
This is analogous to \code{refcat = NA} in \code{\link{categorical}}
models. As the resulting model is only weakly identified, informative
priors on the \code{theta} parameters are strongly recommended (see the
examples in \code{\link{brmsformula}}).}

\item{gr}{Optional name (as a character string) of a grouping variable in the
data. If specified, the mixture is computed over the levels of that variable
rather than over individual observations, that is, all observations that
belong to the same group are assumed to stem from the same (unknown) mixture
component. This is useful, for example, when entire participants (rather than
single trials) are thought to belong to different latent classes. By default
(\code{NULL}), the mixture is computed at the observation level as usual.
See Details for restrictions of group-level mixtures.}
}
\value{
An object of class \code{mixfamily}.
}
\description{
Set up a finite mixture family for use in \pkg{brms}.
}
\details{
Most families supported by \pkg{brms} can be used to form mixtures. The
response variable has to be valid for all components of the mixture family.
Currently, the number of mixture components has to be specified by the user.
It is not yet possible to estimate the number of mixture components from the
data.

Ordering intercepts in mixtures of ordinal families is not possible as each
family has itself a set of vector of intercepts (i.e. ordinal thresholds).
Instead, \pkg{brms} will fix the vector of intercepts across components in
ordinal mixtures, if desired, so that users can try to identify the mixture
model via selective inclusion of predictors.

For most mixture models, you may want to specify priors on the
population-level intercepts via \code{\link{set_prior}} to improve
posterior inference. In addition, it is sometimes necessary to set \code{init = 0}
in the call to \code{\link{brm}} to allow chains to initialize properly.

By default, the mixture is computed at the level of individual observations,
that is, each observation may stem from a different mixture component. Via
argument \code{gr}, the mixture can instead be computed over the levels of a
grouping variable so that all observations within a group share the same
(unknown) component. This yields a different likelihood in which the mixing
is marginalized once per group rather than once per observation. Group-level
mixtures currently require the mixing proportions to be either estimated,
fixed, or predicted by covariates that are constant within each group. They
are not supported in combination with within-chain threading, censoring,
truncation, observation weights, or multivariate models. For group-level
mixtures, \code{log_lik}, \code{loo}, and \code{waic} are computed per group
(i.e. leave-one-group-out cross-validation), while \code{kfold} and the
refinements \code{reloo}, \code{loo_moment_match}, and \code{loo_subsample}
are not available. When predicting from new data, the groups are defined by the
grouping variable within that data set and need not match the groups of the
original data.

For more details on the specification of mixture
models, see \code{\link{brmsformula}}.
}
\examples{
\dontrun{
## simulate some data
set.seed(1234)
dat <- data.frame(
  y = c(rnorm(200), rnorm(100, 6)),
  x = rnorm(300),
  z = sample(0:1, 300, TRUE)
)

## fit a simple normal mixture model
mix <- mixture(gaussian, gaussian)
prior <- c(
  prior(normal(0, 7), Intercept, dpar = mu1),
  prior(normal(5, 7), Intercept, dpar = mu2)
)
fit1 <- brm(bf(y ~ x + z), dat, family = mix,
            prior = prior, chains = 2)
summary(fit1)
pp_check(fit1)

## use different predictors for the components
fit2 <- brm(bf(y ~ 1, mu1 ~ x, mu2 ~ z), dat, family = mix,
            prior = prior, chains = 2)
summary(fit2)

## fix the mixing proportions
fit3 <- brm(bf(y ~ x + z, theta1 = 1, theta2 = 2),
            dat, family = mix, prior = prior,
            init = 0, chains = 2)
summary(fit3)
pp_check(fit3)

## predict the mixing proportions
fit4 <- brm(bf(y ~ x + z, theta2 ~ x),
            dat, family = mix, prior = prior,
            init = 0, chains = 2)
summary(fit4)
pp_check(fit4)

## predict all mixing proportions without a reference category
## (requires informative priors on the theta parameters)
mix_na <- mixture(gaussian, gaussian, refcat = NA)
theta_prior <- c(prior, prior(normal(0, 1), Intercept, dpar = theta1),
                  prior(normal(0, 1), Intercept, dpar = theta2))
fit5 <- brm(bf(y ~ x + z, theta1 ~ x, theta2 ~ x),
            dat, family = mix_na, prior = theta_prior,
            init = 0, chains = 2)
summary(fit5)
pp_check(fit5)

## compare model fit
loo(fit1, fit2, fit3, fit4)

## group-level mixture: whole groups belong to the same component
dat$g <- rep(1:60, each = 5)
fit6 <- brm(bf(y ~ 1), dat, family = mixture(gaussian, gaussian, gr = "g"),
            prior = prior, init = 0, chains = 2)
summary(fit6)
}

}
