% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{mixture}
\alias{mixture}
\title{Mixture Families in \pkg{brms}}
\usage{
mixture(..., flist = NULL, order = NULL)
}
\arguments{
\item{...}{Two or more objects providing a description of the 
response distributions to be combined in the mixture model. 
These can be family functions, calls to family functions or 
character strings naming the families.
For details of supported families see 
\code{\link[brms:brmsfamily]{brmsfamily}}.}

\item{flist}{Optional list of objects, which are treated in the 
same way as objects passed via the \code{...} argument.}

\item{order}{Logical; indicating whether population-level intercepts
of the families should be ordered to identify mixture components.
If \code{NULL} (the default), \code{order} is set to \code{TRUE}
if all families are the same and \code{FALSE} otherwise.}
}
\value{
An object of class \code{mixfamily}.
}
\description{
Set up a mixture family for use in \pkg{brms}.
}
\details{
TODO
}
\examples{
# TODO

}
