% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{ngrps.brmsfit}
\alias{ngrps}
\alias{ngrps.brmsfit}
\title{Number of levels}
\usage{
\method{ngrps}{brmsfit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}.}

\item{...}{Currently ignored.}
}
\value{
A named list containing the number of levels per
  grouping factor
}
\description{
Number of levels of one or more grouping factors
}

