% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{parnames}
\alias{parnames}
\alias{parnames.brmsfit}
\title{Extract Parameter Names}
\usage{
parnames(x, ...)
}
\arguments{
\item{x}{An \R object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector containing the parameter names of the model.
}
\description{
Extract all parameter names of a given model.
}
