% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pp_check}
\alias{pp_check}
\title{temporary pp_check generic function}
\usage{
pp_check(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments}
}
\description{
will be removed as soon as the \pkg{ppcheck} package
has its own generic function
}

