% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.brmsfit}
\alias{print.brmsfit}
\alias{print.brmssummary}
\title{Print a summary for a fitted model represented by a \code{brmsfit} object}
\usage{
\method{print}{brmsfit}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{brmsfit}}

\item{digits}{The number of significant digits for printing out the summary;
defaults to 2. The effective sample size is always rounded to integers.}

\item{...}{Additional arguments that would be passed
to method \code{summary} of \code{brmsfit}.}
}
\description{
Print a summary for a fitted model represented by a \code{brmsfit} object
}
\seealso{
\code{\link{summary.brmsfit}}
}
