% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{R2D2}
\alias{R2D2}
\title{R2D2 Priors in \pkg{brms}}
\usage{
R2D2(mean_R2 = 0.5, prec_R2 = 2, cons_D2 = 0.5, autoscale = TRUE, main = TRUE)
}
\arguments{
\item{mean_R2}{Mean of the Beta prior on the coefficient of determination R^2.}

\item{prec_R2}{Precision of the Beta prior on the coefficient of determination R^2.}

\item{cons_D2}{Concentration vector of the Dirichlet prior on the variance
decomposition parameters. Lower values imply more shrinkage.}

\item{autoscale}{Logical; indicating whether the R2D2
prior should be scaled using the residual standard deviation
\code{sigma} if possible and sensible (defaults to \code{TRUE}).
Autoscaling is not applied for distributional parameters or
when the model does not contain the parameter \code{sigma}.}

\item{main}{Logical; only relevant if the horseshoe prior spans multiple
parameter classes. In this case, only arguments of the single instance for
which \code{main} is \code{TRUE} will be used. See the Examples section
below.}
}
\description{
Function used to set up R2D2 priors for population-level effects in
\pkg{brms}. The function does not evaluate its arguments -- it exists purely
to help set up the model.
}
\details{
Currently, the following classes support the R2D2 prior: \code{b}
  (overall regression coefficients), \code{sds} (SDs of smoothing splines),
  \code{sdgp} (SDs of Gaussian processes), \code{ar} (autoregressive
  coefficients), \code{ma} (moving average coefficients), \code{sderr} (SD of
  latent residuals), \code{sdcar} (SD of spatial CAR structures), \code{sd}
  (SD of varying coefficients).

  Even when the R2D2 prior is applied to multiple parameter classes at once,
  the concentration vector (argument \code{cons_D2}) has to be provided
  jointly in the the one instance of the prior where \code{main = TRUE}. The
  order in which the elements of concentration vector correspond to the
  classes' coefficients is the same as the order of the classes provided
  above.
}
\examples{
set_prior(R2D2(mean_R2 = 0.8, prec_R2 = 10))

}
\references{
Zhang, Y. D., Naughton, B. P., Bondell, H. D., & Reich, B. J. (2020).
  Bayesian regression using a prior on the model fit: The R2-D2 shrinkage
  prior. Journal of the American Statistical Association.
  \url{https://arxiv.org/pdf/1609.00046.pdf}
}
\seealso{
\code{\link{set_prior}}
}
