% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{remove_env_attrs}
\alias{remove_env_attrs}
\title{Recursively remove attached environments from an object}
\usage{
remove_env_attrs(obj)
}
\description{
Recursively remove attached environments from an object
}
\keyword{internal}
