% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{residuals.brmsfit}
\alias{residuals.brmsfit}
\title{Extract Model Residuals from brmsfit Objects}
\usage{
\method{residuals}{brmsfit}(object, type = c("ordinary", "pearson"),
  summary = TRUE, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{type}{The type of the residuals, either \code{ordinary} or \code{pearson}. Currently, the latter
  is only supported for families \code{binomial} and \code{bernoulli}}

\item{summary}{logical. Should summary statistics (i.e. means, sds, and 95\% intervals) be returned
instead of the raw values. Default is \code{TRUE}}

\item{probs}{The percentiles to be computed by the \code{quantile} function. Only used if \code{summary = TRUE}.}

\item{...}{Currently ignored}
}
\value{
Models residuals. If \code{summary = TRUE} this is a N x C matrix and if \code{summary = FALSE}
  a S x N matrix, where S is the number of samples, N is the number of observations,
  and C is equal to \code{length(probs) + 2}.
}
\description{
Extract Model Residuals from brmsfit Objects
}
\details{
Currently, the method does not support \code{categorical} or ordinal models.
}
\examples{
\dontrun{
## fit a model
fit <- brm(rating ~ treat + period + carry + (1|subject), data = inhaler,
           n.cluster = 2)

## extract residuals
res <- residuals(fit, summary = TRUE)
head(res)
}
}

