% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{threading}
\alias{threading}
\title{Threading in Stan}
\usage{
threading(threads = NULL, grainsize = NULL, static = FALSE)
}
\arguments{
\item{threads}{Number of threads to use in within-chain parallelization.}

\item{grainsize}{Number of observations evaluated together in one chunk
on one of the CPUs used for threading.}

\item{static}{Logical. Apply the static (non-adaptive) version of
\code{reduce_sum}? Defaults to \code{FALSE}.}
}
\value{
A \code{brmsthreads} object which can be passed to the
  \code{threads} argument of \code{brm} and related functions.
}
\description{
Use threads for within-chain parallelization in \pkg{Stan} via the \pkg{brms}
interface. Within-chain parallelization is experimental! We recommend its use
only if you are experienced with Stan's \code{reduce_sum} function and have a
slow running model that cannot be sped up by any other means.
}
\examples{
\dontrun{
# this model just serves as an illustration
# threading may not actually speed things up here
fit <- brm(count ~ zAge + zBase * Trt + (1|patient),
           data = epilepsy, family = negbinomial(),
           chains = 1, threads = threading(2, grainsize = 100),
           backend = "cmdstanr")
summary(fit)
}

}
