% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{update.brmsfit}
\alias{update.brmsfit}
\title{Update \pkg{brms} models}
\usage{
\method{update}{brmsfit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{object of class \code{brmsfit}}

\item{newdata}{optional \code{data.frame} to update the model with new data}

\item{...}{other arguments passed to \code{\link[brms:brm]{brm}} such as
   \code{n.iter}, \code{n.warmup}, \code{n.chains}, \code{n.thin},
   \code{n.cluster}, and \code{inits}.}
}
\description{
This method allows to update an existing \code{brmsfit} object with new data
as well as changed configuration of the chains.
}

