% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{update.brmsfit}
\alias{update.brmsfit}
\title{Update \pkg{brms} models}
\usage{
\method{update}{brmsfit}(object, formula., newdata = NULL, ...)
}
\arguments{
\item{object}{object of class \code{brmsfit}}

\item{newdata}{optional \code{data.frame} 
to update the model with new data}

\item{...}{other arguments passed to 
\code{\link[brms:brm]{brm}}}
}
\description{
This method allows to update an existing \code{brmsfit} object
}

