% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{validate_prior}
\alias{validate_prior}
\title{Validate priors supplied by the user}
\usage{
validate_prior(
  prior,
  formula,
  data,
  sample_prior = "no",
  allow_autoscale = TRUE,
  require_nlpar_prior = TRUE,
  ...
)
}
\arguments{
\item{prior}{A character string defining a distribution in \pkg{Stan} language}

\item{formula}{An object of class \code{\link[stats:formula]{formula}},
\code{\link{brmsformula}}, or \code{\link{mvbrmsformula}} (or one that can
be coerced to that classes): A symbolic description of the model to be
fitted. The details of model specification are explained in
\code{\link{brmsformula}}.}

\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{sample_prior}{Indicate if samples from priors should be drawn
additionally to the posterior samples. Options are \code{"no"} (the
default), \code{"yes"}, and \code{"only"}. Among others, these samples can
be used to calculate Bayes factors for point hypotheses via
\code{\link{hypothesis}}. Please note that improper priors are not sampled,
including the default improper priors used by \code{brm}. See
\code{\link{set_prior}} on how to set (proper) priors. Please also note
that prior samples for the overall intercept are not obtained by default
for technical reasons. See \code{\link{brmsformula}} how to obtain prior
samples for the intercept. If \code{sample_prior} is set to \code{"only"},
samples are drawn solely from the priors ignoring the likelihood, which
allows among others to generate samples from the prior predictive
distribution. In this case, all parameters must have proper priors.}

\item{allow_autoscale}{logical, indicating whether autoscaling
by parameter sigma is allowed.}

\item{require_nlpar_prior}{logical; indicating whether priors
are required on coefficients of nlpars.}

\item{...}{Other arguments passed to special prior functions
\code{validate_prior_special} required for priors that are
not natively supported by \pkg{Stan}.}
}
\value{
An object of class \code{brmsprior}.
}
\description{
Validate priors supplied by the user
}
\seealso{
\code{\link{get_prior}}, \code{\link{set_prior}}.
}
