% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R, R/deprecated-methods.R
\name{VarCorr.brmsfit}
\alias{VarCorr.brmsfit}
\alias{VarCorr}
\alias{as.data.frame.brmsVarCorr}
\title{Extract Variance and Correlation Components}
\usage{
\method{VarCorr}{brmsfit}(x, sigma = 1, summary = TRUE, robust = FALSE,
  probs = c(0.025, 0.975), old = FALSE, estimate = "mean", ...)

\method{as.data.frame}{brmsVarCorr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{brmsfit}.}

\item{sigma}{Ignored (included for compatibility with 
\code{\link[nlme:VarCorr]{VarCorr}}).}

\item{summary}{Should summary statistics 
 (i.e. means, sds, and 95\% intervals) be returned
instead of the raw values? Default is \code{TRUE}.}

\item{robust}{If \code{FALSE} (the default) the mean is used as 
the measure of central tendency and the standard deviation as 
the measure of variability. If \code{TRUE}, the median and the 
median absolute deivation (MAD) are applied instead.
Only used if \code{summary} is \code{TRUE}.}

\item{probs}{The percentiles to be computed by the \code{quantile} 
function. Only used if \code{summary} is \code{TRUE}.}

\item{old}{Logical; indicates if the old implementation
of this method (prior to version 1.7.0) should be used.
Defaults to \code{FALSE}.}

\item{estimate}{(Deprecated) A character vector specifying 
which coefficients (e.g., \code{"mean"}, \code{"median"}, 
\code{"sd"}, or \code{"quantile"}) should be calculated 
for the population-level effects. Only used if \code{old}
is \code{TRUE}.}

\item{...}{Further arguments to be passed to the functions 
specified in \code{estimate}}
}
\value{
A list of lists (one per grouping factor), each with
three elements: a matrix containing the standard deviations, 
an array containing the correlation matrix, and an array 
containing the covariance matrix with variances on the diagonial.

If \code{old} is \code{TRUE}, the returned object is of class 
\code{brmsVarCorr}, which can be coerced to a \code{data.frame} 
by using the \code{as.data.frame} method.
}
\description{
This function calculates the estimated standard deviations, 
correlations and covariances of the group-level terms 
in a multilevel model of class \code{brmsfit}. 
For linear models, the residual standard deviations, 
correlations and covariances are also returned.
}
\examples{
\dontrun{
fit <- brm(count ~ log_Age_c + log_Base4_c * Trt_c + (1+Trt_c|visit), 
           data = epilepsy, family = gaussian(), chains = 2)
VarCorr(fit)
}

}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}
