% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit-methods.R
\name{vcov.brmsfit}
\alias{vcov.brmsfit}
\title{Covariance and Correlation Matrix of Fixed Effects}
\usage{
\method{vcov}{brmsfit}(object, correlation = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{brmsfit}}

\item{correlation}{logical; if \code{FALSE} (the default), 
compute the covariance matrix,
if \code{TRUE}, compute the correlation matrix}

\item{...}{Currently ignored}
}
\value{
covariance or correlation matrix of fixed effects parameters
}
\description{
Get a point estimate of the covariance or 
correlation matrix of fixed effects parameters
}
\details{
Estimates are obtained by calculating the maximum likelihood 
  covariances (correlations) of the posterior samples.
}

