% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.generics.R
\name{WAIC}
\alias{WAIC}
\title{Compute the WAIC}
\usage{
WAIC(x, ...)
}
\arguments{
\item{x}{A fitted model object typically of class \code{brmsfit}.}

\item{...}{Optionally more fitted model objects.}
}
\description{
Compute the Watanabe-Akaike Information Criterion based on the posterior likelihood
}
\details{
When comparing models fitted to the same data, the smaller the WAIC, the better the fit.
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}
\references{
Gelman, A., Hwang, J., & Vehtari, A. (2014). Understanding predictive information criteria for Bayesian models.
Statistics and Computing, 24, 997-1016.

Watanabe, S. (2010). Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory.
The Journal of Machine Learning Research, 11, 3571-3594.
}

