% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{WAIC}
\alias{WAIC}
\alias{WAIC.brmsfit}
\title{Compute the WAIC}
\usage{
WAIC(x, ..., compare = TRUE)
}
\arguments{
\item{x}{A fitted model object typically of class \code{brmsfit}.}

\item{...}{Optionally more fitted model objects.}

\item{compare}{A flag indicating if the WAICs of the models should be compared to each other}
}
\value{
If just one object is provided, an object of class \code{ic}.
If multiple objects are provided, an object of class \code{iclist}.
}
\description{
Compute the Watanabe-Akaike Information Criterion based on the posterior likelihood
by using the \pkg{loo} package
}
\details{
When comparing models fitted to the same data, the smaller the WAIC, the better the fit.
}
\examples{
\dontrun{
#model with fixed effects only
fit_i1 <- brm(rating ~ treat + period + carry,
              data = inhaler, family = "gaussian")
WAIC(fit_i1)

#model with an additional random intercept for subjects
fit_i2 <- brm(rating ~ treat + period + carry + (1|subject),
             data = inhaler, family = "gaussian")
#compare both models
WAIC(fit_i1, fit_i2)
}
}
\author{
Paul-Christian Buerkner \email{paul.buerkner@gmail.com}
}
\references{
Vehtari, A., Gelman, A., and Gabry, J. (2015). Efficient implementation of leave-one-out cross-validation and WAIC for evaluating fitted Bayesian models.

Gelman, A., Hwang, J., & Vehtari, A. (2014). Understanding predictive information criteria for Bayesian models.
Statistics and Computing, 24, 997-1016.

Watanabe, S. (2010). Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory.
The Journal of Machine Learning Research, 11, 3571-3594.
}

